### 2025-12-05 0030272
Reduce length of paper
