+++ paper/paper.md
@@ -68,45 +68,45 @@ The Munich Urban Carbon Column network (MUCCnet) consists of 5 EM27/SUN spectrom
 
 We made the following architectural decisions to achieve this higher degree of automation and ease of use:
 
-- **Configuration:** JSON Configuration files are used to specify where to find the input data, which tasks to perform, which arguments to use during the retrieval, and so on. These files are validated against JSON schemas [@jsonschema;@pydantic] representing a strict set of rules of how a configuration file can look. By this, a misconfiguration will be detected immediately, as opposed to only failing when the code tries to access invalid parts of the configuration. The JSON schema is rendered as an API reference in the documentation, enabling us to manage and adequately document a large number of configuration parameters.
-- **Metadata Management:** Metadata (where/when/how each sensor was set up) is provided as JSON files – again, validated against a documented JSON schema. These files can be stored locally or pulled from a GitHub repository, enabling centralized, version-controlled tracking of sensor deployments. We believe all metadata systems used in different EM27/SUN teams can be easily translated into the schema we established with this pipeline. Our metadata schema design largely agrees with the schema proposed in @Zeeman2024.
+- **Configuration:** JSON configuration files are used to specify input/output paths, which date range to retrieve, which retrieval algorithm to use, and so on. These files are validated against JSON schemas [@jsonschema;@pydantic] representing a strict set of rules for their structure and content. A misconfiguration will be detected immediately at program start, as opposed to only failing when the code tries to access invalid parts of the configuration. The JSON schema is rendered as an API reference in the documentation, enabling us to manage and adequately document a large number of configuration parameters.
+- **Metadata Management:** Metadata (where/when/how each sensor was set up) is provided as JSON files – again, validated against a documented JSON schema. These files can be stored locally or pulled from a GitHub repository, enabling centralized, version-controlled tracking of sensor deployments. We believe all metadata systems used in different EM27/SUN teams can be easily translated into the schema this pipeline establishes. Our metadata schema design largely agrees with the schema proposed in @Zeeman2024.
 - **Parallelization and Containerization:** While the PROFFASTpylot parallelizes the individual steps of PROFFAST, the pipeline parallelizes the whole retrieval process by running one isolated retrieval job per sensor per day. Every result folder maintains the structure of an individual PROFFASTpylot run, but also stores all configuration files required to fully reproduce the output in this folder.
 - **Algorithm Support:** While each tagged release of the PROFFASTpylot supports a specific range of PROFFAST versions, the EM27 Retrieval Pipeline keeps support for all implemented retrieval algorithms and versions – the user simply has to specify which algorithm and version to use for a job. Furthermore, we added our own interface for PROFFAST 1 retrievals. The pipeline's architecture is independent of the underlying retrieval algorithm; hence, it could be extended to support GFIT [@Connor2016;@Zeng2021] or other algorithms.
-- **Caching/Implicit Scheduling:** Users do not have to define an explicit queue of jobs, but instead define a date range and a list of sensors to include. The downloading/retrieval/bundling will then decide which jobs to run. I.e., the pipeline only requests atmospheric profiles that have not been downloaded yet, and it only runs retrievals for data with no existing retrieval result (for the same algorithm and atmospheric profile).
[…]
