+++ paper/paper.bib
@@ -7,6 +7,13 @@
   version = {1.15.0},
   date = {2025-02-05},
 }
+@software{pyright,
+  author = {{Microsoft}},
+  title = {Pyright: Static Type Checker for Python},
+  url = {https://github.com/microsoft/pyright},
+  version = {1.1.407},
+  date = {2025-10-25},
+}
 @software{pydantic,
   author = {{Pydantic}},
   title = {Pydantic: Data validation using Python type hints},
@@ -586,6 +593,15 @@ Software},
   author = {Stauber, Josef and Chen, Jia and Klappenbach, Friedrich and Li, Junwei and Luther, Andreas and Makowski, Moritz and Tang, Haoyue and Ponomarev, Nikita and Brunner, Dominik},
   year = {2025},
   month = mar
+}
+@article{Chen2025ICUC,
+  title = {Observational and Modeling Tools for Monitoring Urban Greenhouse Gas Emissions: Results of the ICOS Pilot City Munich},
+  url = {http://dx.doi.org/10.5194/icuc12-721},
+  DOI = {10.5194/icuc12-721},
+  publisher = {Copernicus GmbH},
+  author = {Chen,  Jia and Stauber,  Josef and Li,  Junwei and Aigner,  Patrick and K\"{u}hbacher,  Daniel and Makowski,  Moritz and Luther,  Andreas and Wenzel,  Adrian and Tang,  Haoyue and Hinderer,  Julian and Dietrich,  Florian and Asam,  Christoph},
+  year = {2025},
+  month = may 
 }
 
 % grants
