### 2025-12-21 87bd3f6
Further reduce length of paper
