### 2025-12-21 2150d19
Debug Python version restriction
