+++ pyproject.toml
@@ -42,7 +42,7 @@ classifiers = [
     "Programming Language :: Python :: 3.13",
     "Programming Language :: Python :: 3.14",
 ]
-requires-python = ">=3.11,<=3.14"
+requires-python = ">=3.11,<3.15"
 readme = "README.md"
 license = { text = "GPL-3.0" }
 
@@ -72,9 +72,7 @@ distribution = false
 # using pytest markers from the `pytest-order` library, we run the
 # quick tests before the retrieval tests to fail fast
 [tool.pytest.ini_options]
-filterwarnings = [
-    "ignore::UserWarning",
-]
+filterwarnings = ["ignore::UserWarning"]
 markers = [
     "integration: can only be run on a configured system",
     "quick: can be run in a CI environment, < 30 seconds, everything except running actual retrievals",
@@ -148,9 +146,4 @@ exclude = [
     "dev",
     "docs",
 ]
-strict = [
-    "src",
-    "tests",
-    "run.py",
-    "cli.py",
-]
+strict = ["src", "tests", "run.py", "cli.py"]
