### 2025-12-21 ac911bc
Enable full CI tests again on Python 3.11
