### 2026-01-14 a501f05
Update paper
rename and reorder sections according to new JOSS guidelines
