+++ paper/paper.md
@@ -64,7 +64,19 @@ The EM27/SUN [@Gisi2011;@Hase2016] is an FTIR spectrometer used to perform solar
 
 \vspace{-2mm}
 
-The Munich Urban Carbon Column network (MUCCnet) consists of 5 EM27/SUN spectrometers deployed in and around Munich, Germany [@Heinle2018;@Dietrich2021;@Aigner2023]. Due to its autonomous operation since 2019, our research group generates a significant amount of EM27/SUN data (see \autoref{table:coverage}). Processing this data in an automated and scalable way required us to extend the functionality of the PROFFASTpylot. Under the hood, we use the PROFFASTpylot to ensure consistency with the COCCON community. The EM27 Retrieval Pipeline is complementary to the PROFFASTpylot, but the more EM27/SUN measurement data one has to retrieve, the more benefits come from this higher degree of automation. While not implementing 100% of the flexibility the PROFFASTpylot provides, this pipeline is flexible enough to be used by any EM27/SUN setup, adhering to the COCCON standards.
+The Munich Urban Carbon Column network (MUCCnet) consists of 5 EM27/SUN spectrometers deployed in and around Munich, Germany [@Heinle2018;@Dietrich2021;@Aigner2023]. Due to its autonomous operation since 2019, our research group generates a significant amount of EM27/SUN data. Processing this data in an automated and scalable way required us to extend the functionality of the PROFFASTpylot. Under the hood, we use the PROFFASTpylot to ensure consistency with the COCCON community. The EM27 Retrieval Pipeline is complementary to the PROFFASTpylot, but the more EM27/SUN measurement data one has to retrieve, the more benefits come from this higher degree of automation. While not implementing 100% of the flexibility the PROFFASTpylot provides, this pipeline is flexible enough to be used by any EM27/SUN setup, adhering to the COCCON standards.
+
+# Statement of Need
+
+Human-induced climate change is one of the most pressing issues of our time, as it puts the stability of our Earth system at risk. Our civilization is built on that stability. Anthropogenic GHG emissions, as well as natural sources and sinks of GHGs, have to be quantified to create targeted reduction policies, measure policy effectiveness, and identify new emission sources [@Calvin2023].
+
+The EM27/SUN is widely used to achieve these GHG flux estimations [@Hase2015;@Klappenbach2015;@Chen2016;@Butz2017;@Vogel2019;@Luther2019;@Dietrich2021;@Jones2021;@Luther2022;@Tu2022;@Forstmaier2023;@Doc2025journal;@Stauber2025]. Unlike the HR125 used by TCCON [@tccon], the EM27/SUN can be transported and deployed at different and remote locations without much effort. Since 2016, the instruments of our permanent urban sensor network MUCCnet have been deployed at 3 locations in San Francisco [@Klappenbach2021], 4 in Hamburg [@Forstmaier2023], 4 in Vienna [@Luther2023], 12 in Poland [@Luther2019;@Luther2022], and 11 locations in Munich [@Chen2018;@Dietrich2021].
+
+\autoref{fig:data} shows the XCO2 timeseries of the MUCCnet instruments since September 2019. Running the retrievals for various instrument deployments and keeping track of all deployments over time requires an organizational system that this pipeline provides. To enable a large number of long-term observations with EM27/SUN spectrometers, such as the upcoming sensor networks of COCCON-Spain [@cocconspain] and GEMINI-UK [@geminiuk], an improvement in automation is crucial. The EM27 Retrieval Pipeline addresses this.
+
[…]
