### 2026-01-14 1c64154
Update paper
add AI usage disclaimer
