+++ paper/paper.md
@@ -98,6 +98,10 @@ Furthermore, the pipeline adds features required to produce long-term EM27/SUN d
 
 The pipeline was first established in 2021 [@Rimann2022] and has been used for our entire EM27/SUN dataset since 2022 [@Luther2023;@Klappenbach2024;@Chen2024;@Tang2024;@Loew2025;@Stauber2025;@Chen2025ICUC]. Multiple journal publications using MUCCnet for inverse modeling are currently in preparation. The data generated by MUCCnet and the EM27 Retrieval Pipeline has been published to the ICOS Carbon Portal [@Makowski2023CarbonPortal;@Makowski2024CarbonPortal] and the EVDC Portal [@Loew2025EVDC;@Chen2025EVDCOCOPF10;@Chen2025EVDCOCOPF22] and is a cruicial data input for the GHG modeling approaches developed with the grants listed below.
 
+# AI Usage Disclosure
+
+GitHub Copilot was used for single line code-completions, but not to write any complete functions or modules. We tried using GitHub Copilot Agent once for a small feature in a pull request, but did merge the result due to a lack in quality.
+
 # Acknowledgement of Financial Support
 
 This work is funded by the Horizon 2020 ICOS Cities PAUL Project under [grant no. 101037319](https://cordis.europa.eu/project/id/101037319), the German Federal Ministry of Education and Research (BMBF) project 'Integrated Greenhouse Gas Monitoring System for Germany' (ITMS) under [grant no. 01LK2303B](https://www.fona.de/en/measures/funding-measures/ITMS_en.php), and the ERC Consolidator Grant CoSense4Climate under [grant no. 101089203](https://cordis.europa.eu/project/id/101089203).
