### 2026-01-14 cd0b95a
Update paper
shorten software design a bit
