### 2026-01-14 a8d624d
Fix pytests for Python3.14
Ignore soft deprecations
