+++ tests/repository/test_historic_opus_stat_parsing.py
@@ -13,7 +13,7 @@ def test_historic_opus_stat_parsing() -> None:
     git_status_before = tum_esm_utils.shell.run_shell_command("git status")
     print(git_status_before)
     assert (
-        os.system(
+        os.system(  # pyright: ignore[reportDeprecated]
             f"cd {PROJECT_DIR} && {sys.executable} src/scripts/parse_historic_opus_file_stats.py --force"
         )
         == 0
