### 2026-01-22 80f5fd0
#142 Validate GEOMS writer with PROFFASTpylot implementation (1)
Link relevant equivalent loading functions from pylot repository
