### 2026-01-22 003da54
#142 Validate GEOMS writer with PROFFASTpylot implementation (2)
Extend gases loaded from colsens file from 8 to 10
