### 2026-01-22 05c1a2e
#142 Validate GEOMS writer with PROFFASTpylot implementation (3)
Link more relevant equivalent loading functions from pylot repository
