+++ src/geoms/geoms_api.py
@@ -117,6 +117,8 @@ class GEOMSAPI:
             else:
                 dataset.attrs[key] = np.bytes_(value)
 
+    # https://github.com/coccon/proffastpylot/blob/2.4.1-2/prfpylot/output/hdf_geoms_writer.py#L705
+    # identical to PROFFASTpylot 2.4.1-2
     @staticmethod
     def write_source(hdf5_file: h5py.File) -> None:
         """Source information"""
@@ -132,6 +134,8 @@ class GEOMSAPI:
         )
         GEOMSAPI._write_to_hdf5_file(hdf5_file, variable_name, data, metadata)
 
+    # https://github.com/coccon/proffastpylot/blob/2.4.1-2/prfpylot/output/hdf_geoms_writer.py#L730
+    # identical to PROFFASTpylot 2.4.1-2
     @staticmethod
     def write_datetime(hdf5_file: h5py.File, df: pd.DataFrame) -> None:
         """Datetime information"""
@@ -156,6 +160,8 @@ class GEOMSAPI:
         )
         GEOMSAPI._write_to_hdf5_file(hdf5_file, variable_name, data, metadata)
 
+    # https://github.com/coccon/proffastpylot/blob/2.4.1-2/prfpylot/output/hdf_geoms_writer.py#L767
+    # identical to PROFFASTpylot 2.4.1-2
     @staticmethod
     def write_altitude(hdf5_file: h5py.File, df: pd.DataFrame, pt_df: pd.DataFrame) -> None:
         """Altitude information used in the a-priori profile matrix"""
@@ -181,6 +187,8 @@ class GEOMSAPI:
         )
         GEOMSAPI._write_to_hdf5_file(hdf5_file, variable_name, data, metadata)
 
+    # https://github.com/coccon/proffastpylot/blob/2.4.1-2/prfpylot/output/hdf_geoms_writer.py#L802
+    # identical to PROFFASTpylot 2.4.1-2
     @staticmethod
     def write_solar_angle_zenith(hdf5_file: h5py.File, df: pd.DataFrame) -> None:
         """Solar zenith angle"""
@@ -201,6 +209,8 @@ class GEOMSAPI:
         )
         GEOMSAPI._write_to_hdf5_file(hdf5_file, variable_name, data, metadata)
 
+    # https://github.com/coccon/proffastpylot/blob/2.4.1-2/prfpylot/output/hdf_geoms_writer.py#L832
+    # identical to PROFFASTpylot 2.4.1-2
     @staticmethod
     def write_solar_angle_azimuth(hdf5_file: h5py.File, df: pd.DataFrame) -> None:
         """Solar azimuth angle"""
@@ -225,6 +235,8 @@ class GEOMSAPI:
         )
         GEOMSAPI._write_to_hdf5_file(hdf5_file, variable_name, data, metadata)
 
+    # https://github.com/coccon/proffastpylot/blob/2.4.1-2/prfpylot/output/hdf_geoms_writer.py#L871
+    # identical to PROFFASTpylot 2.4.1-2
     @staticmethod
     def write_instrument_latitude(hdf5_file: h5py.File, df: pd.DataFrame) -> None:
         """Instrument latitude"""
@@ -244,6 +256,8 @@ class GEOMSAPI:
         )
         GEOMSAPI._write_to_hdf5_file(hdf5_file, variable_name, data, metadata)
 
+    # https://github.com/coccon/proffastpylot/blob/2.4.1-2/prfpylot/output/hdf_geoms_writer.py#L900
+    # identical to PROFFASTpylot 2.4.1-2
     @staticmethod
[…]
