### 2026-01-22 cef9193
#142 Validate GEOMS writer with PROFFASTpylot implementation (4)
Make loading of column sensitivity "smart", i.e., detect the gas names
from the file itself rather than using fixed indices
