### 2026-01-22 3b53605
#143 Automatically test GEOMS writer (2)
Add test script
