+++ tests/geoms/test_geoms_writing.py
@@ -1,0 +1,150 @@
+import datetime
+import os
+import pytest
+import src
+import tum_esm_utils
+from ..fixtures import download_sample_data  # pyright: ignore[reportUnusedImport]
+
+PROJECT_DIR = tum_esm_utils.files.rel_to_abs_path("../..")
+INPUT_DATA_DIR = os.path.join(PROJECT_DIR, "data", "testing", "inputs")
+METADATA_DIR = os.path.join(INPUT_DATA_DIR, "metadata")
+BUNDLE_OUTPUT_DIR = os.path.join(PROJECT_DIR, "data", "testing", "bundle", "outputs")
+
+CONFIG = {
+    "version": "1.9",
+    "general": {
+        "metadata": None,
+        "data": {
+            "ground_pressure": {
+                "path": os.path.join(INPUT_DATA_DIR, "data", "ground-pressure"),
+                "file_regex": "^$(SENSOR_ID)$(DATE).*\\.csv$",
+                "separator": ",",
+                "pressure_column": "pressure",
+                "pressure_column_format": "hPa",
+                "date_column": "UTCdate_____",
+                "date_column_format": "%Y-%m-%d",
+                "time_column": "UTCtime_____",
+                "time_column_format": "%H:%M:%S",
+            },
+            "atmospheric_profiles": os.path.join(INPUT_DATA_DIR, "data", "atmospheric-profiles"),
+            "interferograms": os.path.join(INPUT_DATA_DIR, "data", "interferograms"),
+            "results": os.path.join(INPUT_DATA_DIR, "results"),
+        },
+    },
+    "profiles": None,
+    "retrieval": None,
+    "bundles": None,
+    "geoms": {
+        "sensor_ids": ["so", "mc"],
+        "retrieval_algorithms": [
+            "proffast-2.2",
+            "proffast-2.3",
+            "proffast-2.4",
+            "proffast-2.4.1",
+        ],
+        "atmospheric_profile_models": ["GGG2014", "GGG2020"],
+        "from_datetime": "2017-01-01T00:00:00+0000",
+        "to_datetime": "2024-12-31T23:59:59+0000",
+        "parse_dc_timeseries": True,
+        "max_sza": 80,
+        "min_xair": 0.98,
+        "max_xair": 1.02,
+        "conflict_mode": "replace",
+    },
+}
+
+GEOMS_METADATA = {
+    "general": {
+        "network": "FTIR.COCCON",
+        "affiliation": "TUM.ESM",
+        "pressure_sensor_name": "young-61302",
+    },
+    "data": {
+        "discipline": "ATMOSPHERIC.CHEMISTRY;REMOTE.SENSING;GROUNDBASED",
+        "group": "EXPERIMENTAL;PROFILE.STATIONARY",
+        "file_version": 2,
+        "quality": "Station data.",
+        "template": "GEOMS-TE-FTIR-COCCON-001",
+    },
+    "file": {
+        "doi": "10.48477/COCCON.ITMS-B-FTIR.R01",
+        "meta_version": "04R093;CUSTOM",
+        "access": "COCCON",
+        "project_id": "COCCON",
+    },
+    "principle_investigator": {
+        "name": "Chen;Jia",
+        "email": "jia.chen@tum.de",
+        "affiliation": "Technical University of Munich - Professorship of Environmental Sensing and Modeling;TUM.ESM",
+        "address": "Theresienstr. 90;D-80333 Munich;GERMANY",
+    },
+    "data_originator": {
+        "name": "Makowski;Moritz",
+        "email": "moritz.makowski@tum.de",
+        "affiliation": "Technical University of Munich - Professorship of Environmental Sensing and Modeling;TUM.ESM",
[…]
