### 2026-01-22 5625e36
#143 Automatically test GEOMS writer (3)
Make GEOMS bundler accept test arguments instead of loading the from the config folder
