### 2026-01-22 3867abe
#143 Automatically test GEOMS writer (4)
Debug colsens file moving
