### 2026-01-22 608366d
#143 Automatically test GEOMS writer (6)
Finish GEOMS writer tests
