### 2026-01-22 c7b6851
#143 Automatically test GEOMS writer (7)
Debug GEOMS writer
