### 2026-01-22 7c36649
#143 Automatically test GEOMS writer (8)
Debug GEOMS writer
