+++ CITATION.cff
@@ -39,5 +39,5 @@ doi: 10.5281/zenodo.14284968
 license:
   - gpl-3.0-or-later
 type: software
-version: 1.9.1
-date-released: "2025-12-03"
+version: 1.10.0
+date-released: "2026-01-22"
