### 2026-01-22 087fb0b
Refine GEOMS Exporter
make min number of datapoints per day configurable
