### 2026-01-28 1b5b564
Update JOSS paper
Add "State of the Field" section
