+++ paper/paper.bib
@@ -140,6 +140,25 @@ spectra},
   doi = {10.1364/FTS.2009.JMA3},
   abstract = {A network of ground-based, sun-viewing, near-IR, Fourier transform spectrometers has been established to accurately measure atmospheric greenhouse gases such as CO2, CO, N2O, and CH4.},
 }
+@article{Laughner2024,
+  title = {The Total Carbon Column Observing Network’s  GGG2020 data version},
+  volume = {16},
+  ISSN = {1866-3516},
+  url = {http://dx.doi.org/10.5194/essd-16-2197-2024},
+  DOI = {10.5194/essd-16-2197-2024},
+  number = {5},
+  journal = {Earth System Science Data},
+  publisher = {Copernicus GmbH},
+  author = {Laughner,  Joshua L. and Toon,  Geoffrey C. and Mendonca,  Joseph and Petri,  Christof and Roche,  Sébastien and Wunch,  Debra and Blavier,  Jean-Francois and Griffith,  David W. T. and Heikkinen,  Pauli and Keeling,  Ralph F. and Kiel,  Matth\"{a}us and Kivi,  Rigel and Roehl,  Coleen M. and Stephens,  Britton B. and Baier,  Bianca C. and Chen,  Huilin and Choi,  Yonghoon and Deutscher,  Nicholas M. and DiGangi,  Joshua P. and Gross,  Jochen and Herkommer,  Benedikt and Jeseck,  Pascal and Laemmel,  Thomas and Lan,  Xin and McGee,  Erin and McKain,  Kathryn and Miller,  John and Morino,  Isamu and Notholt,  Justus and Ohyama,  Hirofumi and Pollard,  David F. and Rettinger,  Markus and Riris,  Haris and Rousogenous,  Constantina and Sha,  Mahesh Kumar and Shiomi,  Kei and Strong,  Kimberly and Sussmann,  Ralf and Té,  Yao and Velazco,  Voltaire A. and Wofsy,  Steven C. and Zhou,  Minqiang and Wennberg,  Paul O.},
+  year = {2024},
+  month = may,
+  pages = {2197–2260}
+}
+@article{Wunch2015,
+  title = {The Total Carbon Column Observing Network’s GGG2014 data version, Pasadena, California},
+  author = {Wunch, Debra and Toon, Geoffrey C. and Sherlock, Vanessa and Deutscher, Nicholas M. and Liu, C. and Feist, D. G. and Wennberg, Paul O.},
+  year = {2015},
+}
 
 % em27
 @article{Gisi2011,
