### 2026-01-28 #9548 @dostuffthatmatters
Dear @arfon, dear @kthyng,

sure, no worries. We added the section with a bit more comparison to the alternatives. Without the header, the paper is now at 1191 words. We hope that is ok, given the amount of citations.

Best regards,
Moritz Oliviera Makowski
