### 2026-02-17 #9548 @dostuffthatmatters
Hi @kthyng, hi @arfon,

I hope you’re both doing well — is there anything we can do on our end to help move the pre-review along? I believe, the project meets all of the pre-review checklist items.

Best,
Moritz Oliveira Makowski
