### 2026-03-17 #9548 @kthyng
I'm going to put the new verison of my checklist here:

- [x]  [Not a minor utility](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance)

New submission criteria since journal scope changes in January 2026 (more details in [submission docs](https://joss.readthedocs.io/en/latest/submitting.html#pre-review-screening-criteria)):
- [x] Public development history ≥ 6 months
- [x] Demonstrated research impact
- [x] Open source practices
- [x] Iterative development
- [x] Check for borderline cases
