### 2026-03-23 #9548 @dostuffthatmatters
Hi @rwegener2

great thank you so much for editing this! @patrickjaigner has a conflict of interest because we have been coleagues in the same lab for 4 years now. @feldl is a coauthor of this publication.

I believe the following reviewers from the JOSS reviewers database are all qualified and non-conflicting to review this (this got lost a bit in the thread above):

- `falconstryker` (https://github.com/falconstryker, https://reviewers.joss.theoj.org/reviewers/17729)
- `blsqr` (https://github.com/blsqr, https://reviewers.joss.theoj.org/reviewers/1611)
- `cbueth` (https://github.com/cbueth, https://reviewers.joss.theoj.org/reviewers/4003)
- `DS4Ag` (https://github.com/DS4Ag, https://reviewers.joss.theoj.org/reviewers/13043)
- `cjerzak` (https://github.com/cjerzak, https://reviewers.joss.theoj.org/reviewers/17994)

All the best,
Moritz Oliveira Makowski
