### 2026-04-14 #10413 @editorialbot
[REVIEW]: EM27 Retrieval Pipeline: Automated EM27/SUN Data Processing
**Submitting author:** @dostuffthatmatters (<a href="http://orcid.org/0000-0002-2948-2993">Moritz Makowski</a>)
**Repository:** https://github.com/tum-esm/em27-retrieval-pipeline
**Branch with paper.md** (empty if default branch):
**Version:** 1.10.0
