### 2026-04-14 #10413 @rwegener2
👋🏼 @dostuffthatmatters @DS4Ag, @dorchard this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
