### 2026-04-28 tum-esm/em27-retrieval-pipeline#147 @DS4Ag
JOSS - General checks
### 1. License

You are supposed to fill the template with your software information, for instance:

```
<program>  Copyright (C) <year>  <name of author>
This program comes with ABSOLUTELY NO WARRANTY;
```

---

### 2. Contribution and authorship

Please review your list of authors and follow the JOSS guidelines:
https://joss.readthedocs.io/en/latest/submitting.html#authorship

In particular:
- Explain the role of authors who do not have commits
- Explain why contributors with commits are not listed as authors

| Name in paper                 | Software repository contributor | Number of commits | Explanation |
|-----------------------------|--------------------------------|-------------------|-------------|
| Moritz Oliveira Makowski     | @dostuffthatmatters            | 1408              |             |
| Frank Hase                   |                                |                   |             |
| Friedrich Klappenbach        |                                |                   |             |
| Andreas Luther               |                                |                   |             |
| Lena Feld                    |                                |                   |             |
| Marlon Müller                | @MarlonMueller                 | 130               |             |
| Vyas Giridharan              | @vyasgiridhar                  | 26                |             |
| Catherine Fait              |                                |                   |             |
| Jia Chen                     |                                |                   |             |
|                              | @undernagruzez                 | 29                |             |
|                              | @cfleur                        | 7                 |             |
