### 2026-04-28 tum-esm/em27-retrieval-pipeline#148 @DS4Ag
JOSS - Documentation
1.  Automated tests
   - Please improve test coverage for files that are below 80%.
   - It would be helpful to integrate a coverage reporting tool such as Codecov to provide a centralized graphical view of test coverage.

2. Does the software have any specific prerequisites? If that is the case, include it in the README section and in the documentation.

3.  In the usage section of the documentation, you state: “We highly recommend using PDM — a Python package manager — over pip.” However, the installation instructions use pip. Please include two versions of the installation instructions: one using PDM and the other using pip.

4. The documentation could be further strengthened. For instance, if the software is designed to be used as a data extraction pipeline, it may be helpful to clearly describe how to use each step of the pipeline.

5. I suggest including one or more examples in the documentation that show, step by step, how to use the software for specific situations.

6. Do not include the repository README on the documentation page. For instance, it currently contains a link that refers back to the documentation page itself.

7. Please include Community Guidelines.

8. In the README, you mention:

   > “We retrieve a lot of EM27/SUN data, produced mainly by MUCCnet (Dietrich et al., 2021), and have used this pipeline since the end of 2021.”

   Does this mean the software is intended only to retrieve that specific dataset? Can it be used by other users to retrieve their own data collected using the EM27/SUN spectrometer?

9. The README file could also be expanded. For example, it might be helpful to include:
   - A clear statement of the problem the software is intended to solve
   - A summary of its main functionalities
   - A simple usage example
