### 2026-04-28 tum-esm/em27-retrieval-pipeline#149 @DS4Ag
JOSS - Statement of Need
In general, I notice that citations are sometimes overused for the same concept. For instance, “The EM27/SUN (Gisi et al., 2011; Hase et al., 2016)”. Is there a difference in what is described by the first and second references? Do they refer to different versions of the spectrometer? If not, it may be sufficient to cite only the original paper. Over-citation can make the text harder to follow.

**L10.** Please include the full name of Fourier-transform infrared (FTIR).

**L12.** I assume ESA refers to the European Space Agency.

**L12–13.** What is the reason for including three citations for COCCON? If each reference contributes something different, please clarify their respective contributions. Otherwise, it may be sufficient to cite the original paper describing its establishment. The citation may be better placed where COCCON is first described.

**L13.** Avoid using two parentheses side by side. Instead, you can update as follows:
Network (COCCON; Alberti et al., 2021; …)

**L17–18.** Consider citing only the original paper, unless later works introduce important changes that need to be highlighted.

**L17–19.** The term PROFFAST is repeated multiple times. Consider rephrasing for clarity and conciseness.

**L20–22.** Consider citing only the first paper describing MUCCnet.

**Figure 1.**
- Figures should be referenced in the text before they appear.
- The text in Figure 1 is difficult to read due to its small size. JOSS does not seem to provide an option to enlarge figures, so improving readability would be helpful.

**L22.** When you refer to “our research group,” does this correspond to MUCCnet?

**L23–24.** Is this need specific to the MUCCnet research group, or is it more broadly applicable to other research groups?

**L33.** What does GFIT stand for? Please define it and cite the original reference.

**L35.** "GGG2014/GGG2020 (J. L. Laughner et al., 2024; Wunch et al., 2015)." Do these references correspond to different versions (GGG2014 vs. GGG2020)? If so, please clarify.

**L36.** Instead of:
(TCCON) (Toon et al., 2009)
Use:
(TCCON; Toon et al., 2009)

**L47–48.** Since there is a large number of citations here, consider summarizing them in a table that highlights differences (e.g., type of research, context, or contribution).

**L50.** Is HR125 another spectrometer? Are HR125 and EM27/SUN the only available options? It may be sufficient to indicate that EM27/SUN is portable, instead of comparing it with the HR125.

**Figure 2.**
I am not sure whether Figure 2 adds information relevant to the software presented in this paper (EM27 Retrieval Pipeline). It may be worth reconsidering its inclusion.

**L89–90.** Please fix the typo:
(using PyTest (Krekel et al., 2024))

**L93–94.** Please cite GGG2014, GGG2020, and Ginput individually. For example:
Ginput server (J. Laughner et al., 2021)

**L98–99.**  You mention: “For the full feature set of the pipeline, please refer to its GitHub repository.”
However, the README does not appear to contain a complete description of the full feature set.

**L100.** For consistency, consider including the documentation website URL, as was done for the GitHub repository.

**L101–102.** Instead of the current Figure 2, consider including a version timeline that highlights the main features introduced in each release, especially since previous versions are mentioned.

**L104.** “The pipeline was first established in 2021 (Rißmann et al., 2022).”
 - Does this mean that version 1.0.0 of the EM27 Retrieval Pipeline was published in 2022?
- Does the citation (Rißmann et al., 2022) refer to a previous publication of the software itself, or to a study that used the software?

**L105.**  “has been used for our entire EM27/SUN dataset since 2022.”
It may be more informative to include details about the dataset, such as the number of variables, records, or sites. Also, does “our entire dataset” refer to all data collected by MUCCnet across all locations?

**L104–107**  The manuscript lists several publications to demonstrate the use of the pipeline. However, the contribution of the software to these studies is not clearly specified.

To better align with JOSS guidelines, it would be helpful to briefly clarify how the software enabled these publications (e.g., data processing, analysis pipeline, or integration into workflows). This would strengthen the research impact statement by making the evidence of impact more explicit and specific.

**L107–108.** The statement “Multiple journal publications … are currently in preparation” is forward-looking and does not provide verifiable evidence of impact.

**L110.**  Do the citations correspond to different versions of the EVDC portal? Why are multiple citations needed here?
