### 2026-04-28 #10413 @DS4Ag
Hello @dostuffthatmatters

My initial comments are on the following GitHub issues:
- https://github.com/tum-esm/em27-retrieval-pipeline/issues/147
- https://github.com/tum-esm/em27-retrieval-pipeline/issues/148
- https://github.com/tum-esm/em27-retrieval-pipeline/issues/149
