### 2026-05-11 002d780
Debug pytests
Do not test spectra from mc instrument for PROFFAST 2.4.1 due to stronger filtering criteria - this will be fixed in a future version.
