### 2026-05-11 8c2d874
Add issue templates
