### 2026-05-11 2d7537a
Add issue templates
