+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,0 +1,5 @@
+blank_issues_enabled: true
+contact_links:
+  - name: Documentation
+    url: https://em27-retrieval-pipeline.netlify.app/
+    about: Please read the documentation before submitting an issue.
