### 2026-05-11 c9a4bc2
#150 Add full example directory (2)
Remove old test data download from pytests
