+++ tests/fixtures.py
@@ -6,33 +6,12 @@ import tum_esm_utils
 import src
 
 _PROJECT_DIR = tum_esm_utils.files.rel_to_abs_path("..")
-_TEST_DATA_DIR = os.path.join(_PROJECT_DIR, "data", "testing", "inputs", "data")
 
 
 @pytest.fixture
 def wrap_test_with_mainlock() -> Generator[None, None, None]:
     with src.utils.semaphores.with_automation_lock():
         yield
-
-
-@pytest.fixture(scope="session")
-def download_sample_data() -> Generator[None, None, None]:
-    """Download sample data from a GitHub release because it is too big for the repository.
-    Could also be done with a Git LFS but this way it is easier to share for other purposes."""
-
-    for label in ["interferograms", "atmospheric-profiles", "ground-pressure"]:
-        tum_esm_utils.code.download_github_release_asset(
-            repository="tum-esm/em27-retrieval-pipeline",
-            asset_name=f"em27-retrieval-pipeline-tests-data-20250723-{label}.tar.gz",
-            dst_dir=_TEST_DATA_DIR,
-            final_name=f"{label}.tar.gz",
-            access_token=os.getenv("GITHUB_API_TOKEN", None),
-        )
-        tum_esm_utils.shell.run_shell_command(
-            f"tar -xzf {label}.tar.gz", working_directory=_TEST_DATA_DIR
-        )
-
-    yield
 
 
 @pytest.fixture(scope="session")
