### 2026-05-11 a9aa60b
#150 Add full example directory (2)
Remove old test data download from pytests
