### 2026-05-11 3207ae0
#150 Add full example directory (3)
Point tests to example data
remove old test data directories and bundling script
