### 2026-05-11 59807c0
#150 Add full example directory (5)
Point all test cases to newly structure test data
