+++ tests/retrieval/test_retrieval_queue.py
@@ -70,7 +70,7 @@ def test_retrieval_queue(
 
     # target config at test data
     config.general.data.ground_pressure.path.root = os.path.join(
-        PROJECT_DIR, "data", "testing", "inputs", "data", "ground-pressure"
+        PROJECT_DIR, "example", "data", "inputs", "ground-pressure"
     )
     config.general.data.ground_pressure.pressure_column = "BaroYoung"
     config.general.data.ground_pressure.pressure_column_format = "hPa"
@@ -83,13 +83,13 @@ def test_retrieval_queue(
     config.general.data.ground_pressure.time_column_format = "%H:%M:%S"
 
     config.general.data.interferograms.root = os.path.join(
-        PROJECT_DIR, "data", "testing", "inputs", "data", "interferograms"
+        PROJECT_DIR, "example", "data", "inputs", "interferograms"
     )
     config.general.data.atmospheric_profiles.root = os.path.join(
-        PROJECT_DIR, "data", "testing", "inputs", "data", "atmospheric-profiles"
+        PROJECT_DIR, "example", "data", "inputs", "atmospheric-profiles"
     )
     config.general.data.results.root = os.path.join(
-        PROJECT_DIR, "data", "testing", "container", "outputs"
+        PROJECT_DIR, "data", "testing", "outputs", "individual-results"
     )
 
     # test case 1
