+++ src/scripts/parse_historic_opus_file_stats.py
@@ -10,8 +10,8 @@ sys.path.append(tum_esm_utils.files.rel_to_abs_path("../.."))
 from src import utils
 
 # TODO: set these paths
-IFG_PATH = tum_esm_utils.files.rel_to_abs_path("../../data/testing/inputs/data/interferograms")
-RESULTS_PATH = tum_esm_utils.files.rel_to_abs_path("../../data/testing/inputs/results")
+IFG_PATH = tum_esm_utils.files.rel_to_abs_path("../../example/data/inputs/interferograms")
+RESULTS_PATH = tum_esm_utils.files.rel_to_abs_path("../../data/testing/inputs/individual-results")
 
 if __name__ == "__main__":
     # call as "s.py allowedsid1 allowedsid2 ..."
