### 2026-05-11 2825969
#150 Add full example directory (6)
Allow relative paths in config
