### 2026-05-11 6ce29eb
#150 Add full example directory (7)
Add function to run retrieval in foreground, refine config template
