### 2026-05-11 925dcec
#150 Add full example directory (8)
Add script to run example
