### 2026-05-11 bf6dad9
#150 Add full example directory (9)
Also run GEOMS export in example
