### 2026-05-11 a997871
#150 Add full example directory (10)
Store geoms outputs in a central place
