### 2026-05-11 6a6e5f3
#150 Add full example directory (11)
Test full example in pytest in Ci tests
