### 2026-05-11 decad47
#150 Add full example directory (11)
Test full example in pytest in Ci tests
