### 2026-05-11 82431ee
#150 Add full example directory (12)
Exclude things from code coverage report that are very hard to test manually: mainly profile downloader and exceptions
