### 2026-05-11 2848870
#150 Add full example directory (13)
Debug tests
