### 2026-05-11 24af46b
#150 Add full example directory (13)
Debug tests
