### 2026-05-11 9b2325e
#150 Add full example directory (14)
Further improve test coverage by removing old outputs before running the tests
