### 2026-05-11 dab6eac
#150 Add full example directory (14)
Further improve test coverage by removing old outputs before running the tests
