+++ tests/retrieval/test_container.py
@@ -75,7 +75,9 @@ SENSOR_DATA_CONTEXTS = [
 @pytest.fixture(scope="session")
 def provide_container_factory_for_ci_tests(
     provide_config_template: src.types.Config,
-) -> Generator[src.retrieval.dispatching.container_factory.ContainerFactory, None, None]:
+) -> Generator[
+    src.retrieval.dispatching.container_factory.ContainerFactory, None, None
+]:  # pragma: no cover
     logger = src.retrieval.utils.logger.Logger(
         "pytest",
         write_to_file=False,
@@ -89,7 +91,9 @@ def provide_container_factory_for_ci_tests(
 @pytest.fixture(scope="session")
 def provide_container_factory_for_complete_tests(
     provide_config_template: src.types.Config,
-) -> Generator[src.retrieval.dispatching.container_factory.ContainerFactory, None, None]:
+) -> Generator[
+    src.retrieval.dispatching.container_factory.ContainerFactory, None, None
+]:  # pragma: no cover
     logger = src.retrieval.utils.logger.Logger(
         "pytest",
         write_to_file=False,
@@ -102,7 +106,7 @@ def provide_container_factory_for_complete_tests(
 
 @pytest.mark.order(3)
 @pytest.mark.quick
-def test_sdc_covers_the_full_day() -> None:
+def test_sdc_covers_the_full_day() -> None:  # pragma: no cover
     for sdc in SENSOR_DATA_CONTEXTS:
         assert src.utils.functions.sdc_covers_the_full_day(sdc)
 
@@ -115,7 +119,7 @@ def test_container_lifecycle_ci(
     remove_temporary_retrieval_data: None,
     provide_config_template: src.types.Config,
     provide_container_factory_for_ci_tests: src.retrieval.dispatching.container_factory.ContainerFactory,
-) -> None:
+) -> None:  # pragma: no cover
     config = provide_config_template
     container_factory = provide_container_factory_for_ci_tests
 
@@ -152,7 +156,7 @@ def test_container_lifecycle_complete(
     remove_temporary_retrieval_data: None,
     provide_config_template: src.types.Config,
     provide_container_factory_for_complete_tests: src.retrieval.dispatching.container_factory.ContainerFactory,
-) -> None:
+) -> None:  # pragma: no cover
     config = provide_config_template
     container_factory = provide_container_factory_for_complete_tests
 
@@ -230,7 +234,7 @@ def test_container_lifecycle_complete(
 
 def run_session(
     session: src.types.RetrievalSession, config: src.types.Config, only_run_mock_retrieval: bool
-) -> None:
+) -> None:  # pragma: no cover
     src.retrieval.session.process_session.run(
         config,
         session,
@@ -249,7 +253,7 @@ def _generate_job_list() -> list[
         src.types.AtmosphericProfileModel,
         em27_metadata.types.SensorDataContext,
     ]
-]:
+]:  # pragma: no cover
     src.retrieval.utils.retrieval_status.RetrievalStatusList.reset()
 
     pending_jobs: list[
@@ -283,7 +287,7 @@ def _generate_job_list() -> list[
     return pending_jobs
 
 
-def _point_config_to_test_data(config: src.types.Config) -> None:
+def _point_config_to_test_data(config: src.types.Config) -> None:  # pragma: no cover
     config.general.data.ground_pressure.path.root = os.path.join(
         PROJECT_DIR, "example", "data", "inputs", "ground-pressure"
     )
@@ -312,7 +316,7 @@ def _assert_output_correctness(
     retrieval_algorithm: src.types.RetrievalAlgorithm,
     atmospheric_profile_model: src.types.AtmosphericProfileModel,
[…]
