+++ README.md
@@ -2,7 +2,7 @@
 
 <img alt="GitHub License" src="https://img.shields.io/github/license/tum-esm/em27-retrieval-pipeline?style=flat&label=License&labelColor=%230f172a&color=%23fef08a&cacheSeconds=60" className="inline p-0 m-px mt-4"/> <img alt="GitHub Tag" src="https://img.shields.io/github/v/tag/tum-esm/em27-retrieval-pipeline?sort=semver&style=flat&label=Latest%20Pipeline%20Version&color=%23fef08a&cacheSeconds=60&labelColor=%230f172a" className="inline p-0 m-px mt-4"/> <br/> <img alt="Static Badge" src="https://img.shields.io/badge/Proffast%201.0%20%7C%202.2%20%7C%202.3%20%7C%202.4-whydoineedthis?style=flat&label=Retrieval%20Algorithms&labelColor=%230f172a&color=%2399f6e4&cacheSeconds=60" className="inline p-0 m-px "/> <br/> <img alt="Static Badge" src="https://img.shields.io/badge/GGG2014%20%7C%20GGG2020%20-%20whydoineedthis?style=flat&label=Atmospheric%20Profile%20Models&labelColor=%230f172a&color=%2399f6e4&cacheSeconds=60" className="inline p-0 m-px "/> <br/> [<img alt="Static Badge" src="https://img.shields.io/badge/DOI-10.5281/zenodo.14284968-%23bae6fd?labelColor=%230f172a"/>](https://doi.org/10.5281/zenodo.14284968)
 
-We retrieve a lot of EM27/SUN data, produced mainly by [MUCCnet (Dietrich et al., 2021)](https://doi.org/10.5194/amt-14-1111-2021), and have used this pipeline since end of 2021.
+We retrieve a lot of EM27/SUN data, due to running [MUCCnet (Dietrich et al., 2021)](https://doi.org/10.5194/amt-14-1111-2021) in an automated setup, and have used this pipeline since end of 2021.
 
 This codebase provides an automated data pipeline for [Proffast 1 and 2.X](https://www.coccon.kit.edu). Under the hood, it uses the [Proffast Pylot](https://github.com/coccon/proffastpylot) to interact with Proffast 2 and an in-house connector to interact with Proffast 1. Whenever using this pipeline for Proffast retrievals, please make sure to also cite Proffast and the Proffast Pylot (for Proffast 2.X retrievals).
 
@@ -14,3 +14,69 @@ This codebase provides an automated data pipeline for [Proffast 1 and 2.X](https
 
 ⚙️ Many of our projects (including this pipeline) use functionality from the [`tum-esm-utils` package](https://github.com/tum-esm/utils).<br/>
 🤖 Our EM27/SUN systems are running autonomously with the help of [Pyra](https://github.com/tum-esm/pyra).
+
+
+## EM27 Retrieval Pipeline vs. PROFFASTpylot
+
+This pipeline does not intend to replace the Proffast Pylot by any means. It uses most of the Pylot's functionality and provides an automation layer on top of that. You can use Proffast directly for full flexibility, the Pylot for some flexibility, or this pipeline, which is more opinionated than the Pylot but highly automated. Furthermore, this pipeline will continue to include support for the different retrieval algorithms.
+
+We decided to include a copy of the Python codebase of the Proffast Pylot inside this repository so we can have less complexity due to Git Submodules or on-demand downloads. Since the Proffast Pylot only supports Proffast 2.X, we have added our own wrapper for Proffast 1 in this pipeline.
+
+## What does this pipeline do?
+
+The pipeline consists of three building blocks that are required to retrieve EM27 data:
+
+1. Downloading vertical profiles from the Caltech FTP server:<br/>
[…]
