### 2026-05-11 6407080
Refine readme and update sync script
