+++ tests/repository/test_full_example.py
@@ -5,7 +5,7 @@ import tum_esm_utils
 PROJECT_DIR = tum_esm_utils.files.rel_to_abs_path("../../")
 
 
-@pytest.mark.order(1)
+@pytest.mark.order(4)
 @pytest.mark.ci
 def test_full_example() -> None:
     exit_code = os.system(f"cd {PROJECT_DIR}/example && bash run.sh")  # pyright: ignore[reportDeprecated]
