### 2026-05-12 522f688
Do not run example tests in CI due to runtime and installation
