+++ .github/workflows/test.yaml
@@ -19,6 +19,8 @@ jobs:
     steps:
       - name: Check out repository
         uses: actions/checkout@v4
+        with:
+          lfs: true
       - name: Set up Python ${{ matrix.python_version }}
         uses: actions/setup-python@v5
         with:
