### 2026-05-12 0073d5a
Enable example tests in CI
