### 2026-05-12 fc4079a
Enable example tests in CI
