+++ .github/workflows/test.yaml
@@ -42,16 +42,15 @@ jobs:
       - name: Install dependencies
         run: pip install ".[dev]"
       
-      - name: Load cached sample data and proffast code
+      - name: Load cached proffast code
         uses: actions/cache@v4
         with:
           path: |
-            data/testing/container/em27-retrieval-pipeline-test-inputs-1.0.0.tar.gz
             src/retrieval/algorithms/proffast-1.0/main/prf
             src/retrieval/algorithms/proffast-2.2/main/prf
             src/retrieval/algorithms/proffast-2.3/main/prf
             src/retrieval/algorithms/proffast-2.4/main/prf
-          key: v1.0.0-ci-cache
+          key: v1.11.0-ci-cache
 
       # Run quick tests for all python versions except 3.11
       - name: Run quick tests
