### 2026-06-07 6166dd0
Update architecture plot in JOSS paper
Make font bigger, reduce text, reposition on the second page
