### 2026-06-07 6b54c50
Extend comments in example directory
