### 2026-06-07 a2400b7
Clarify JOSS paper sentence about config validation
