+++ paper/paper.md
@@ -84,7 +84,7 @@ The EM27/SUN is widely used to achieve these GHG flux estimations [@Hase2015;@Kl
 
 We made the following architectural decisions to achieve this higher degree of automation and ease of use:
 
-- **Configuration:** JSON configuration files are used to specify input/output paths, retrieval algorithm, start/end dates, and so on. These files are validated against JSON schemas [@jsonschema;@pydantic] representing a strict set of rules for their structure and content. A misconfiguration will be detected at program start, instead of failing only once the code tries to use an invalid configuration parameter. The JSON schema is rendered as an API reference in the documentation, enabling us to adequately document a large number of configuration parameters.
+- **Configuration:** JSON configuration files are used to specify input/output paths, retrieval algorithm, start/end dates, and so on. These files are validated against JSON schemas [@jsonschema;@pydantic] representing a strict set of rules for their structure and content. A misconfiguration will be detected at program start, as opposed to encountering a runtime exception when the code tries to use a dynamically typed dictionary object that contains invalid configuration parameters. The JSON schema is rendered as an API reference in the documentation, enabling us to adequately document a large number of configuration parameters.
 - **Metadata Management:** Metadata (where/when/how each sensor was set up) is provided as JSON files as well. These files can be stored locally or pulled from a GitHub repository, enabling centralized, version-controlled tracking of sensor deployments. Metadata systems used in different EM27/SUN teams can be easily translated into this pipeline's schema, which largely agrees with the schema proposed in @Zeeman2024.
 - **Parallelization and Containerization:** While the PROFFASTpylot parallelizes the individual steps of PROFFAST, the pipeline parallelizes the whole retrieval process by running one isolated retrieval job per instrument per day. Every result folder maintains the structure of an individual PROFFASTpylot run, but also contains all configuration files required to reproduce it.
 - **Algorithm Support:** While each tagged release of the PROFFASTpylot supports a specific range of PROFFAST versions, the EM27 Retrieval Pipeline keeps support for all implemented retrieval algorithms and versions.
