### 2026-06-07 2b18c9d
Add pytest case "metadata" and clarify docs on which tests to run for metadata
