### 2026-06-07 89a55c8
Add pytest case "metadata" and clarify docs on which tests to run for metadata
