+++ pyproject.toml
@@ -75,6 +75,7 @@ distribution = false
 filterwarnings = ["ignore::UserWarning"]
 markers = [
     "integration: can only be run on a configured system",
+    "metadata: can only be run on a configured system, only tests the connection to the metadata",
     "quick: can be run in a CI environment, < 30 seconds, everything except running actual retrievals",
     "ci: can be run in a CI environment, ~ 1 - 2 min, runs the complete tests but without running the actual retrieval algorithms",
     "complete: can be run in a CI environment, ~ 3 - 10 minutes, runs three sample days from different instruments and locations for every retrieval algorithm and every atmospheric profile",
