### 2026-06-21 9fec757
Refine how to call the ifg corruption detection
In case the binary needs to be recompiled, it now gives it a second try with a fresh recompiled binary
