### 2026-06-21 3f44563
Improve CI test coverage
