### 2026-06-21 e132369
Improve CI test coverage
