### 2026-06-21 fe19c32
Fix CI tests
