+++ src/retrieval/session/move_ifg_files.py
@@ -1,6 +1,5 @@
 import json
 import os
-import subprocess
 from typing import Optional
 
 import tum_esm_utils
