+++ tests/integration/test_profiles_connection.py
@@ -3,6 +3,7 @@ import ftplib
 
 import src
 
+
 @pytest.mark.order(3)
 @pytest.mark.integration
 def test_profiles_connection() -> None:
