### 2026-06-21 ee9c941
Further improve test coverage
