+++ tests/integration/test_config.py
@@ -2,6 +2,7 @@ import pytest
 
 import src
 
+
 @pytest.mark.order(2)
 @pytest.mark.integration
 def test_config() -> None:
