+++ src/retrieval/dispatching/container_factory.py
@@ -157,14 +157,13 @@ class ContainerFactory:
 
     def remove_all_containers(self, include_unknown: bool = False) -> None:
         """Remove all containers."""
+        for container in self.containers:
+            shutil.rmtree(container.container_path)  # pragma: no cover
         if include_unknown:
             for d in os.listdir(self.container_dir):
                 subdir = os.path.join(self.container_dir, d)
-                if os.path.isdir(subdir):  # pragma: no cover
-                    shutil.rmtree(subdir)
-        else:
-            for container in self.containers:  # pragma: no cover
-                shutil.rmtree(container.container_path)
+                if os.path.isdir(subdir):
+                    shutil.rmtree(subdir)  # pragma: no cover
         self.containers = []
         self.label_generator = tum_esm_utils.text.RandomLabelGenerator()
 
@@ -182,8 +181,8 @@ class ContainerFactory:
         else:
             _print("Downloading Proffast 1.0 code")
             for f in [ZIPFILE_NAME, "2021-03-08_prf96-EM27-fast"]:
-                if os.path.exists(os.path.join(ROOT_DIR, f)):  # pragma: no cover
-                    os.remove(os.path.join(ROOT_DIR, f))
+                if os.path.exists(os.path.join(ROOT_DIR, f)):
+                    os.remove(os.path.join(ROOT_DIR, f))  # pragma: no cover
 
             _print("Downloading")
             tum_esm_utils.shell.run_shell_command(
@@ -251,12 +250,12 @@ class ContainerFactory:
         ROOT_DIR = os.path.join(_RETRIEVAL_CODE_DIR, "proffast-2.2", "main")
 
         # DOWNLOAD PROFFAST 2.2 code if it doesn't exist yet
-        if os.path.exists(os.path.join(ROOT_DIR, "prf")):  # pragma: no cover
-            _print("Proffast 2.2 has already been downloaded")
+        if os.path.exists(os.path.join(ROOT_DIR, "prf")):
+            _print("Proffast 2.2 has already been downloaded")  # pragma: no cover
         else:
             _print("Downloading Proffast 2.2 code")
-            if os.path.exists(os.path.join(ROOT_DIR, ZIPFILE_NAME)):  # pragma: no cover
-                os.remove(os.path.join(ROOT_DIR, ZIPFILE_NAME))
+            if os.path.exists(os.path.join(ROOT_DIR, ZIPFILE_NAME)):
+                os.remove(os.path.join(ROOT_DIR, ZIPFILE_NAME))  # pragma: no cover
 
             _print("Downloading")
             tum_esm_utils.shell.run_shell_command(
@@ -292,12 +291,12 @@ class ContainerFactory:
         ROOT_DIR = os.path.join(_RETRIEVAL_CODE_DIR, "proffast-2.3", "main")
 
         # DOWNLOAD PROFFAST 2.3 code if it doesn't exist yet
-        if os.path.exists(os.path.join(ROOT_DIR, "prf")):  # pragma: no cover
-            _print("Proffast 2.3 has already been downloaded")
+        if os.path.exists(os.path.join(ROOT_DIR, "prf")):
+            _print("Proffast 2.3 has already been downloaded")  # pragma: no cover
         else:
             _print("Downloading Proffast 2.3 code")
-            if os.path.exists(os.path.join(ROOT_DIR, ZIPFILE_NAME)):  # pragma: no cover
-                os.remove(os.path.join(ROOT_DIR, ZIPFILE_NAME))
+            if os.path.exists(os.path.join(ROOT_DIR, ZIPFILE_NAME)):
[…]
