### 2026-06-21 78d672e
Run all CI tests for all python versions
