+++ .github/workflows/test.yaml
@@ -34,7 +34,6 @@ jobs:
           cache: "pip"
       
       - name: Install system dependencies
-        if: ${{ matrix.python_version == '3.11' }}
         run: |
           sudo apt-get update
           sudo apt-get install -y gfortran unzip
@@ -42,29 +41,20 @@ jobs:
       - name: Install dependencies
         run: pip install ".[dev]"
       
-      - name: Load cached proffast code
-        uses: actions/cache@v4
-        with:
-          path: |
-            src/retrieval/algorithms/proffast-1.0/main/prf
-            src/retrieval/algorithms/proffast-2.2/main/prf
-            src/retrieval/algorithms/proffast-2.3/main/prf
-            src/retrieval/algorithms/proffast-2.4/main/prf
-          key: v1.11.0-ci-cache
+      # - name: Load cached proffast code
+      #   uses: actions/cache@v4
+      #   with:
+      #     path: |
+      #       src/retrieval/algorithms/proffast-1.0/main/prf
+      #       src/retrieval/algorithms/proffast-2.2/main/prf
+      #       src/retrieval/algorithms/proffast-2.3/main/prf
+      #       src/retrieval/algorithms/proffast-2.4/main/prf
+      #     key: v1.11.0-ci-cache
 
-      # Run quick tests for all python versions except 3.11
-      - name: Run quick tests
-        if: ${{ matrix.python_version != '3.11' }}
-        env:
-          GITHUB_API_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: pytest -m "quick" --cov=src tests/
-
-      # Run all tests for python 3.11
       - name: Run quick and ci tests
-        if: ${{ matrix.python_version == '3.11' }}
         env:
           GITHUB_API_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: pytest -m "quick or ci" --capture=no --cov=src tests/
+        run: pytest -m "quick or ci" --capture=no tests/
 
       - name: Upload retrieval logs
         if: failure()
