### 2026-06-21 9d2794d
Reduce printouts during tests
