+++ .github/pull_request_template.md
@@ -1,0 +1,14 @@
+## Summary
+
+Describe what changed and why.
+
+## Related Issues
+
+- [ ] Closes #...
+
+## Checklist
+
+- [ ] I added or updated tests where appropriate
+- [ ] I updated user-facing documentation where appropriate
+- [ ] I described any breaking changes
+- [ ] I followed the Code of Conduct and contribution guidelines
