### 2026-06-21 f9da464
Apply ruff linting to codebase
