+++ src/geoms/main.py
@@ -5,31 +5,33 @@ which is licensed under the GNU General Public License version 3. The authors of
 original code are Lena Feld, Benedikt Herkommer, Darko Dubravica affiliated with the
 Karlsruhe Institut of Technology (KIT)."""
 
-from typing import Literal, Optional
 import datetime
 import os
 import re
+import sys
+from typing import Literal, Optional
+
 import em27_metadata
-import polars as pl
-import numpy as np
 import h5py  # pyright: ignore[reportMissingTypeStubs]
+import numpy as np
+import polars as pl
 import tqdm
 import tum_esm_utils
-import sys
 
 sys.path.append(tum_esm_utils.files.rel_to_abs_path("../../.."))
 
 import src
+
+from .geoms_api import GEOMSAPI  # type: ignore
 from .utils import (
-    load_comb_invparms_df,
-    get_ils_form_preprocess_inp,
-    load_pt_file,
-    load_vmr_file,
-    load_interpolated_column_sensitivity_file,
     calculate_column_uncertainty,
     geoms_times_to_datetime,
+    get_ils_form_preprocess_inp,
+    load_comb_invparms_df,
+    load_interpolated_column_sensitivity_file,
+    load_pt_file,
+    load_vmr_file,
 )
-from .geoms_api import GEOMSAPI  # type: ignore
 
 
 # fmt: off
@@ -44,16 +46,16 @@ def generate_geoms_file(
     atmospheric_profile_model: src.types.AtmosphericProfileModel,
 ) -> tuple[Optional[str], str]:
     """Generate a GEOMS file within a given results directory.
-    
+        
     Returns: filepath and status message."""
-    
+
     about_json = tum_esm_utils.files.load_json_file(os.path.join(results_dir, "about.json"))
-    
+
     sensor_id: Optional[str] = None
     serial_number: Optional[int] = None
     location: Optional[em27_metadata.types.LocationMetadata] = None
     date: Optional[datetime.date] = None
-    
+
     if "ctx" in about_json["session"]:
         sensor_id = about_json["session"]["ctx"]["sensor_id"]
         serial_number = int(about_json["session"]["ctx"]["serial_number"])
@@ -68,9 +70,9 @@ def generate_geoms_file(
         date = datetime.date.fromisoformat(about_json["session"]["date"])
         from_dt = datetime.datetime.combine(date, datetime.time(0, 0, 0), tzinfo=datetime.timezone.utc)
         to_dt = datetime.datetime.combine(date, datetime.time(23, 59, 59), tzinfo=datetime.timezone.utc)
-    
+
     assert sensor_id is not None, "Sensor ID could not be determined"
-    assert from_dt.date() == to_dt.date(), f"Something is wrong"
+    assert from_dt.date() == to_dt.date(), "Something is wrong"
 
     # determine calibration factors
     from_dt_calibration_factors_index = calibration_factors.get_index(sensor_id, from_dt)
@@ -94,7 +96,7 @@ def generate_geoms_file(
     )
     if len(pl_df) < geoms_config.min_datapoints_per_day:  # pragma: no cover
         return None, f"Not enough data (less than {geoms_config.min_datapoints_per_day} datapoints)"
-    
+
     # determine filename
     start_stop_times = geoms_times_to_datetime([
         np.round((pl_df["JulianDate"].min() - 2451544.5) * 86400.0),  # type: ignore# type: ignore
@@ -102,7 +104,7 @@ def generate_geoms_file(
     ])
     data_from_dt: datetime.datetime = start_stop_times[0]
     data_to_dt: datetime.datetime = start_stop_times[1]
-    
+
[…]
