+++ src/utils/functions.py
@@ -1,8 +1,9 @@
 import datetime
 import os
 from typing import Any, Optional
-import polars as pl
+
 import em27_metadata
+import polars as pl
 import tum_esm_utils
 
 try:
