### 2026-06-21 e772e00
Apply ruff linting to codebase
