+++ src/bundle/load_results.py
@@ -1,8 +1,10 @@
-from typing import Any, Optional
 import datetime
 import os
+from typing import Any, Optional
+
 import polars as pl
 import tum_esm_utils
+
 import src
 
 
