### 2026-06-21 c87f122
Run ruff formatter and linter in CI tests
