+++ .github/workflows/test.yaml
@@ -51,6 +51,9 @@ jobs:
       #       src/retrieval/algorithms/proffast-2.4/main/prf
       #     key: v1.11.0-ci-cache
 
+      - name: Run ruff formatter and linter
+        run: ruff format --check . && ruff check .
+
       - name: Run quick and ci tests
         env:
           GITHUB_API_TOKEN: ${{ secrets.GITHUB_TOKEN }}
