### 2026-06-21 0dcf1fe
Link contributors guide in README
