### 2026-06-21 47a4e39
Debug noon pressure computation around the UTC dateline
