+++ tests/retrieval/test_pressure_averaging.py
@@ -14,12 +14,22 @@ _GROUND_PRESSURE_DIR = tum_esm_utils.files.rel_to_abs_path(
 @pytest.mark.quick
 def test_solar_noon_computation() -> None:
     start_date = datetime.date(2000, 1, 1)
+    rng = random.Random(0)
     for _ in range(100):
-        date = start_date + datetime.timedelta(days=random.randint(0, 365 * 30))
-        lat = random.uniform(-60, 60)
-        lon = random.uniform(-180, 180)
+        date = start_date + datetime.timedelta(days=rng.randint(0, 365 * 30))
+        lat = rng.uniform(-60, 60)
+        lon = rng.uniform(-180, 180)
         solar_noon = src.retrieval.utils.pressure_averaging.compute_solar_noon_time(lat, lon, date)
-        assert solar_noon.date() == date
+        local_solar_noon = solar_noon + datetime.timedelta(hours=lon / 15)
+        assert local_solar_noon.date() == date
+        assert abs(
+            local_solar_noon
+            - datetime.datetime.combine(
+                date,
+                datetime.time(hour=12),
+                tzinfo=datetime.timezone.utc,
+            )
+        ) < datetime.timedelta(minutes=20)
 
 
 @pytest.mark.order(3)
