### 2026-06-22 7fab8af
Debug results moving
