### 2026-07-05 99726f2
Fix some outdated links and typos
