+++ paper/paper.md
@@ -100,7 +100,7 @@ For the full feature set of the pipeline, please refer to its GitHub repository 
 
 # Research Impact Statement
 
-The pipeline was first established in 2021 [@Rimann2022] and has been used since 2022 to process our entire EM27/SUN dataset [@Luther2023;@Klappenbach2024;@Chen2024;@Tang2024;@Loew2025;@Stauber2025;@Chen2025ICUC]. The data generated by MUCCnet and the EM27 Retrieval Pipeline has been published to the ICOS Carbon Portal [@Makowski2023CarbonPortal;@Makowski2024CarbonPortal] and the EVDC Portal [@Loew2025EVDC;@Chen2025EVDCOCOPF10;@Chen2025EVDCOCOPF22] and is a cruicial data input for the GHG modeling approaches developed with the grants listed below.
+The pipeline was first established in 2021 [@Rimann2022] and has been used since 2022 to process our entire EM27/SUN dataset [@Luther2023;@Klappenbach2024;@Chen2024;@Tang2024;@Loew2025;@Stauber2025;@Chen2025ICUC]. The data generated by MUCCnet and the EM27 Retrieval Pipeline has been published to the ICOS Carbon Portal [@Makowski2023CarbonPortal;@Makowski2024CarbonPortal] and the EVDC Portal [@Loew2025EVDC;@Chen2025EVDCOCOPF10;@Chen2025EVDCOCOPF22] and is a crucial data input for the GHG modeling approaches developed with the grants listed below.
 
 # AI Usage Disclosure
 
