+++ README.md
@@ -15,7 +15,6 @@ This codebase provides an automated data pipeline for [Proffast 1 and 2.X](https
 ⚙️ Many of our projects (including this pipeline) use functionality from the [`tum-esm-utils` package](https://github.com/tum-esm/utils).<br/>
 🤖 Our EM27/SUN systems are running autonomously with the help of [Pyra](https://github.com/tum-esm/pyra).
 
-
 ## EM27 Retrieval Pipeline vs. PROFFASTpylot
 
 This pipeline does not intend to replace the Proffast Pylot by any means. It uses most of the Pylot's functionality and provides an automation layer on top of that. You can use Proffast directly for full flexibility, the Pylot for some flexibility, or this pipeline, which is more opinionated than the Pylot but highly automated. Furthermore, this pipeline will continue to include support for the different retrieval algorithms.
@@ -57,7 +56,7 @@ The data flow from input to merged outputs:
 
 **Comprehensive logs and output data management:** It will store failed and succeeded containers. The output is the same as with the Pylot but also contains all config files the pipeline used to run this container and logs generated by the container.
 
-**Bundling of retrieval results:** The raw retrieval outputs will be distributed over hundres or thousands of folders and files. The bundling script of this pipeline will merge these outputs into one file per station, retrieval algorithm, and atmospheric profile. This way, you can easily access the data for further processing. ([Read about it here](/guides/directories#bundles)).
+**Bundling of retrieval results:** The raw retrieval outputs will be distributed over hundreds or thousands of folders and files. The bundling script of this pipeline will merge these outputs into one file per station, retrieval algorithm, and atmospheric profile. This way, you can easily access the data for further processing. ([Read about it here](/guides/directories#bundles)).
 
 **Documentation and complete API reference:** hosted at [em27-retrieval-pipeline.netlify.app](https://em27-retrieval-pipeline.netlify.app/)
 
