### 2026-07-05 ffe32cb
Expand caption of figure 2
