### 2026-07-05 3a56d75
Update COCCON website links in README, docs and paper
