+++ README.md
@@ -4,7 +4,7 @@
 
 We retrieve a lot of EM27/SUN data, due to running [MUCCnet (Dietrich et al., 2021)](https://doi.org/10.5194/amt-14-1111-2021) in an automated setup, and have used this pipeline since end of 2021.
 
-This codebase provides an automated data pipeline for [Proffast 1 and 2.X](https://www.coccon.kit.edu). Under the hood, it uses the [Proffast Pylot](https://github.com/coccon/proffastpylot) to interact with Proffast 2 and an in-house connector to interact with Proffast 1. Whenever using this pipeline for Proffast retrievals, please make sure to also cite Proffast and the Proffast Pylot (for Proffast 2.X retrievals).
+This codebase provides an automated data pipeline for [Proffast 1 and 2.X](https://www.coccon.kit.edu/). Under the hood, it uses the [Proffast Pylot](https://github.com/coccon/proffastpylot) to interact with Proffast 2 and an in-house connector to interact with Proffast 1. Whenever using this pipeline for Proffast retrievals, please make sure to also cite Proffast and the Proffast Pylot (for Proffast 2.X retrievals).
 
 📚 Read the documentation at [em27-retrieval-pipeline.netlify.app](https://em27-retrieval-pipeline.netlify.app).<br/>
 💾 Get the source code at [github.com/tum-esm/em27-retrieval-pipeline](https://github.com/tum-esm/em27-retrieval-pipeline).<br/>
