### 2026-07-05 44641fc
Update COCCON website links in README, docs and paper
