+++ paper/paper.bib
@@ -75,7 +75,7 @@
 }
 @url{Hase2023,
   title = {{COCCON} {Data} {Processing}},
-  url = {https://www.imk-asf.kit.edu/english/3225.php},
+  url = {https://www.coccon.kit.edu/english/3225.php},
   language = {en-gb},
   urldate = {2023-05-16},
   author = {Hase, Frank},
