### 2026-08-11 0f00923
Update docstrings
