### 2026-08-30 f55efa6
Run quick and integration tests during example run
