### 2026-08-30 6a26775
#151 Make logs dir configurable
