+++ cli.py
@@ -42,7 +42,7 @@ def _check_config_validity() -> None:
 @retrieval_command_group.command(
     name="start",
     short_help="Start Retrieval Process in the Background",
-    help="Start the retrieval as a background process. Prevents spawning multiple processes. The logs and the current processing queue from this process can be found at `data/logs/retrieval`.",
+    help="Start the retrieval as a background process. Prevents spawning multiple processes. The logs and the current processing queue from this process can be found at `$ERP_LOGS_DIR/retrieval` (default: `data/logs/retrieval`).",
 )
 def start() -> None:
     _check_config_validity()
@@ -55,7 +55,7 @@ def start() -> None:
 @retrieval_command_group.command(
     name="run",
     short_help="Start Retrieval Process in the Foreground",
-    help="Run the retrieval. Blocking. The logs and the current processing queue from this process can be found at `data/logs/retrieval`.",
+    help="Run the retrieval. Blocking. The logs and the current processing queue from this process can be found at `$ERP_LOGS_DIR/retrieval` (default: `data/logs/retrieval`).",
 )
 def run() -> None:
     import src
@@ -63,7 +63,8 @@ def run() -> None:
     with src.utils.semaphores.with_automation_lock():
         click.echo("Starting retrieval process in the foreground. Press Ctrl+C to stop.")
         click.echo(
-            "The logs and the current processing queue from this process can be found at `data/logs/retrieval`."
+            "The logs and the current processing queue from this process can be found at "
+            "$ERP_LOGS_DIR/retrieval (default: data/logs/retrieval)."
         )
         src.retrieval.main.run()
 
@@ -71,7 +72,7 @@ def run() -> None:
 @retrieval_command_group.command(
     name="is-running",
     short_help="Check If Retrieval Is Running",
-    help="Checks whether the retrieval background process is running. The logs and the current processing queue from this process can be found at `data/logs/retrieval`.",
+    help="Checks whether the retrieval background process is running. The logs and the current processing queue from this process can be found at `$ERP_LOGS_DIR/retrieval` (default: `data/logs/retrieval`).",
 )
 def is_running() -> None:
     # no config check because this does not require a config
@@ -118,7 +119,7 @@ def watch(
 @retrieval_command_group.command(
     name="stop",
     short_help="Stop Retrieval Process",
-    help="Stop the retrieval background process. The logs and the current processing queue from this process can be found at `logs/retrieval`.",
+    help="Stop the retrieval background process. The logs and the current processing queue from this process can be found at `$ERP_LOGS_DIR/retrieval` (default: `data/logs/retrieval`).",
 )
 def stop() -> None:
     # no config check so that the process can always be terminated
