### 2026-08-30 b828847
#154 Migrate configuration from JSON to TOML (1)
Add new config template
