### 2026-08-30 715e7e4
#154 Migrate configuration from JSON to TOML (2)
Add some vscode specific project settings
