### 2026-09-03 #10413 @dostuffthatmatters
Hi @dorchard,

Apologies for the super late reply! I moved to a different town and wanted to thoroughly address your usability experience.

Thank you for the direct and helpful comment! We want to make the tool as usable as possible, not only for people who are very experienced with the EM27/SUN retrieval workflow. We have been working on some significant usability improvements and will ping you by next week once we have implemented them. One small comment: in the review, I believe, you have used the released v1.10 documentation, but the unreleased `dev-1.11` code. I will release the new 1.11 version with the new changes so that the code and docs are in sync again, and you can retry the much-improved version. Any further changes can of course then be a v1.12.

Some of the usability improvements:

- Merging the `em27-metadata` library into the same repository to avoid the confusing two-repo setup
- Making the local metadata the default recommended way and adding a separate documentation page on how to use the remote metadata feature
- Adding the quick and integration test calls to the example directory
- Running the full example directory in CI as well so that we can be sure, this works plug-and-play
- Moving the whole configuration from JSON to TOML to make it easier to read and document - we made sure that any old locally present config JSON files will be automatically converted to TOML so that this release does not break any existing installation when upgrading
- Significantly improving the schema documentation on the docs website after the migration to TOML: By now, we have added so many options that the nested API reference style [used right now](https://em27-retrieval-pipeline.netlify.app/api-reference/configuration/) does not work nicely anymore. The new configuration reference will be section-wise and will use the same renderer as one of our other projects: [see example](https://tum-esm.github.io/retro/reference/config)

Best regards,
Moritz Oliveira Makowski
