### 2026-09-07 da4b976
Fix typos in the documentation
