### 2026-09-07 970f742
#154 Migrate configuration from JSON to TOML (3)
Move all old schema definitions to a common archival file
