+++ src/types/old_schemas.py
@@ -1,0 +1,813 @@
+from __future__ import annotations
+from typing import Any, Literal, Optional
+import os
+import datetime
+import dotenv
+import pydantic
+import tum_esm_utils
+from .basic_types import AtmosphericProfileModel, RetrievalAlgorithm
+
+
+# CONFIG
+
+
+class MetadataConfig(pydantic.BaseModel):
+    """GitHub repository where the location data is stored."""
+
+    model_config = pydantic.ConfigDict(extra="forbid")
+
+    github_repository: str = pydantic.Field(
+        ...,
+        pattern=r"^[a-z0-9-_]+/[a-z0-9-_]+$",
+        description="GitHub repository name, e.g. `my-org/my-repo`.",
+    )
+    access_token: Optional[str] = pydantic.Field(
+        default=None,
+        min_length=1,
+        description="GitHub access token with read access to the repository, only required if the repository is private.",
+    )
+
+
+class GroundPressureConfig(pydantic.BaseModel):
+    """Format of the ground pressure files. We support any text file that stores one data point per row and separates the columns with a comma, space, or tab, i.e. CSV, TSV, or space-separated files. Using the `file_regex` field, you specify which files to consider for a given sensor id and date.
+
+    You have to specify the columns that contain the date and time of the data. There is three options to specify this - the CLI will complain if you configure none or more than one of these options:
+
+    * One column with a datetime string -> configure `datetime_column` and `datetime_column_format`
+    * Two columns, one with the date and one with the time -> configure `date_column`, `date_column_format`, `time_column`, and `time_column_format`
+    * One column with a unix timestamp -> configure `unix_timestamp_column` and `unix_timestamp_column_format`"""
+
+    model_config = pydantic.ConfigDict(extra="forbid")
+
+    # where to find the files
+    path: tum_esm_utils.validators.StrictDirectoryPath = pydantic.Field(
+        ...,
+        description="Directory path to ground pressure files.",
+    )
+
+    # how to find and parse the files
+    file_regex: str = pydantic.Field(
+        ...,
+        min_length=1,
+        description=(
+            "A regex string to match the ground pressure file names. In this string, you can use the placeholders `$(SENSOR_ID)`, `$(YYYY)`, `$(YY)`, `$(MM)`, and `$(DD)` to make this regex target a certain station and date. The placeholder `$(DATE)` is a shortcut for `$(YYYY)$(MM)$(DD)`."
+        ),
+        examples=[
+            "^$(DATE).tsv$",
+            "^$(SENSOR_ID)_$(DATE).dat$",
+            "^ground-pressure-$(SENSOR_ID)-$(YYYY)-$(MM)-$(DD).csv$",
+        ],
+    )
+    separator: str = pydantic.Field(
+        ...,
+        description="Separator used in the ground pressure files. Only needed and used if the file format is `text`.",
+        min_length=1,
+        max_length=1,
+        examples=[",", "\t", " ", ";"],
+    )
+
+    # one datetime column
+    datetime_column: Optional[str] = pydantic.Field(
+        default=None,
+        description="Column name in the ground pressure files that contains the datetime.",
+        examples=["datetime", "dt", "utc-datetime"],
+    )
+    datetime_column_format: Optional[str] = pydantic.Field(
+        default=None,
+        description="Format of the datetime column in the ground pressure files.",
+        examples=["%Y-%m-%dT%H:%M:%S"],
+    )
+
+    # two columns for date and time
+    date_column: Optional[str] = pydantic.Field(
+        default=None,
+        description="Column name in the ground pressure files that contains the date.",
+        examples=["date", "d", "utc-date"],
+    )
+    date_column_format: Optional[str] = pydantic.Field(
[…]
