### 2026-09-07 fe3925d
#154 Migrate configuration from JSON to TOML (4)
Add new config schema definition
