+++ src/types/config.py
@@ -12,159 +12,226 @@ from .basic_types import AtmosphericProfileModel, RetrievalAlgorithm
 
 
 class MetadataConfig(pydantic.BaseModel):
-    """GitHub repository where the location data is stored."""
+    """How and where to get the metadata from."""
 
     model_config = pydantic.ConfigDict(extra="forbid")
 
-    github_repository: str = pydantic.Field(
+    source: Literal["local", "github"] = pydantic.Field(
         ...,
+        description="Where to source the metadata from. If `local`, it will use `config/metadata.toml`. If `github`, it will download the metadata from the GitHub repository specified in the `github_repository` field.",
+    )
+    github_repository: Optional[str] = pydantic.Field(
+        default=None,
         pattern=r"^[a-z0-9-_]+/[a-z0-9-_]+$",
         description="GitHub repository name, e.g. `my-org/my-repo`.",
     )
-    access_token: Optional[str] = pydantic.Field(
+    github_access_token: Optional[str] = pydantic.Field(
         default=None,
         min_length=1,
         description="GitHub access token with read access to the repository, only required if the repository is private.",
     )
 
+    @pydantic.model_validator(mode="after")
+    def check_github_repository(self) -> MetadataConfig:
+        if self.source == "github" and (self.github_repository is None):
+            raise ValueError("You have to set `github_repository` if you set `source` to `github`")
+        return self
 
-class GroundPressureConfig(pydantic.BaseModel):
-    """Format of the ground pressure files. We support any text file that stores one data point per row and separates the columns with a comma, space, or tab, i.e. CSV, TSV, or space-separated files. Using the `file_regex` field, you specify which files to consider for a given sensor id and date.
 
-    You have to specify the columns that contain the date and time of the data. There is three options to specify this - the CLI will complain if you configure none or more than one of these options:
+class DataConfigs:
+    class AtmosphericProfiles(pydantic.BaseModel):
+        """Where to find the atmospheric profile files."""
 
-    * One column with a datetime string -> configure `datetime_column` and `datetime_column_format`
-    * Two columns, one with the date and one with the time -> configure `date_column`, `date_column_format`, `time_column`, and `time_column_format`
-    * One column with a unix timestamp -> configure `unix_timestamp_column` and `unix_timestamp_column_format`"""
+        path: tum_esm_utils.validators.StrictDirectoryPath = pydantic.Field(
+            ...,
+            description="Directory path to atmospheric profile files.",
+        )
 
-    model_config = pydantic.ConfigDict(extra="forbid")
+        # before the validation, if the path is relative, make it absolute based on the cwd
+        @pydantic.model_validator(mode="before")
+        def _make_path_absolute(cls, values: Any) -> Any:
+            assert isinstance(values["path"], str)
+            if not os.path.isabs(values["path"]):
+                values["path"] = os.path.abspath(values["path"])
+            return values
 
-    # where to find the files
-    path: tum_esm_utils.validators.StrictDirectoryPath = pydantic.Field(
-        ...,
-        description="Directory path to ground pressure files.",
-    )
+    class GroundPressure(pydantic.BaseModel):
+        """Format of the ground pressure files. We support any text file that stores one data point per row and separates the columns with a comma, space, or tab, i.e. CSV, TSV, or space-separated files. Using the `file_regex` field, you specify which files to consider for a given sensor id and date.
 
-    # how to find and parse the files
-    file_regex: str = pydantic.Field(
-        ...,
-        min_length=1,
-        description=(
[…]
