### 2026-09-09 6afb0d6
#154 Migrate configuration from JSON to TOML (6)
Add new config template and refine config schema
