### 2026-09-09 49f95c3
#154 Migrate configuration from JSON to TOML (7)
Add new GEOMS config schema
