### 2026-09-09 b42a324
#154 Migrate configuration from JSON to TOML (8)
Add new GEOMS config template
