+++ config/geoms_config.template.toml
@@ -1,0 +1,75 @@
+# METADATA
+
+[metadata.general]
+network = "FTIR.COCCON"
+affiliation = "TUM.ESM"
+pressure_sensor_name = "vaisala-PTB330"
+
+[metadata.data]
+discipline = "ATMOSPHERIC.CHEMISTRY;REMOTE.SENSING;GROUNDBASED"
+group = "EXPERIMENTAL;PROFILE.STATIONARY"
+file_version = 1
+quality = "Station data."
+template = "GEOMS-TE-FTIR-COCCON-001"
+
+[metadata.file]
+doi = ""
+meta_version = "04R088;CUSTOM"
+access = "COCCON"
+project_id = "COCCON"
+
+[metadata.principle_investigator]
+name = "Chen;Jia"
+email = "jia.chen@tum.de"
+affiliation = "Technical University of Munich - Professorship of Environmental Sensing and Modeling;TUM.ESM"
+address = "Theresienstr. 90;D-80333 Munich;GERMANY"
+
+[metadata.data_originator]
+name = "Makowski;Moritz"
+email = "moritz.makowski@tum.de"
+affiliation = "Technical University of Munich - Professorship of Environmental Sensing and Modeling;TUM.ESM"
+address = "Theresienstr. 90;D-80333 Munich;GERMANY"
+
+[metadata.data_submitter]
+name = "Makowski;Moritz"
+email = "moritz.makowski@tum.de"
+affiliation = "Technical University of Munich - Professorship of Environmental Sensing and Modeling;TUM.ESM"
+address = "Theresienstr. 90;D-80333 Munich;GERMANY"
+
+[metadata.locations]
+TUM_I = "MUNICH.TUM"
+FEL = "MUNICH.FELDKIRCHEN"
+GRAE = "MUNICH.GRAEFELFING"
+OBE = "MUNICH.OBERSCHLEISSHEIM"
+TAU = "MUNICH.TAUFKIRCHEN"
+DLR_2 = "MUNICH.DLR_2"
+DLR_3 = "MUNICH.DLR_3"
+
+# CALIBRATION FACTORS
+
+[[calibration_factors]]
+sensor_id = "ma"
+valid_from_datetime = "2019-01-01T00:00:00Z"
+valid_to_datetime = "2019-12-31T23:59:59Z"
+xco2 = 1.0000
+xch4 = 1.0000
+xco = 1.0000
+xh2o = 1.0000
+
+[[calibration_factors]]
+sensor_id = "mb"
+valid_from_datetime = "2019-01-01T00:00:00Z"
+valid_to_datetime = "2019-12-31T23:59:59Z"
+xco2 = 1.0005
+xch4 = 1.0005
+xco = 1.0005
+xh2o = 1.0005
+
+[[calibration_factors]]
+sensor_id = "ma"
+valid_from_datetime = "2020-01-01T00:00:00Z"
+valid_to_datetime = "2024-12-31T23:59:59Z"
+xco2 = 1.0010
+xch4 = 1.0010
+xco = 1.0010
+xh2o = 1.0010
