### 2026-09-09 45d2d27
#154 Migrate configuration from JSON to TOML (9)
Use new schemas in all steps of the pipeline (didnt update tests yet)
