+++ src/retrieval/utils/job_queue.py
@@ -7,10 +7,8 @@ from src import types
 
 
 class RetrievalJob(pydantic.BaseModel):
-    retrieval_algorithm: types.RetrievalAlgorithm
-    atmospheric_profile_model: types.AtmosphericProfileModel
     sensor_data_context: em27_metadata.types.SensorDataContext
-    job_settings: types.config.RetrievalJobSettingsConfig
+    job_config: types.RetrievalSubConfigs.Job
 
 
 class RetrievalJobQueue:
@@ -20,17 +18,13 @@ class RetrievalJobQueue:
 
     def push(
         self,
-        retrieval_algorithm: types.RetrievalAlgorithm,
-        atmospheric_profile_model: types.AtmosphericProfileModel,
         sensor_data_context: em27_metadata.types.SensorDataContext,
-        job_settings: types.config.RetrievalJobSettingsConfig,
+        job_config: types.RetrievalSubConfigs.Job,
     ) -> None:
         self.queue.append(
             RetrievalJob(
-                retrieval_algorithm=retrieval_algorithm,
-                atmospheric_profile_model=atmospheric_profile_model,
                 sensor_data_context=sensor_data_context,
-                job_settings=job_settings,
+                job_config=job_config,
             )
         )
 
