+++ src/geoms/main.py
@@ -37,9 +37,8 @@ from .utils import (
 # fmt: off
 def generate_geoms_file(
     results_dir: str,
-    geoms_metadata: src.types.GEOMSMetadata,
-    calibration_factors: src.types.CalibrationFactorsList,
     geoms_config: src.types.GEOMSConfig,
+    geoms_metadata: src.types.GEOMSMetadata,
     from_datetime: datetime.datetime,
     to_datetime: datetime.datetime,
     retrieval_algorithm: src.types.RetrievalAlgorithm,
@@ -75,8 +74,8 @@ def generate_geoms_file(
     assert from_dt.date() == to_dt.date(), "Something is wrong"
 
     # determine calibration factors
-    from_dt_calibration_factors_index = calibration_factors.get_index(sensor_id, from_dt)
-    to_dt_calibration_factors_index = calibration_factors.get_index(sensor_id, to_dt)
+    from_dt_calibration_factors_index = geoms_metadata.calibration_factors.get_index(sensor_id, from_dt)
+    to_dt_calibration_factors_index = geoms_metadata.calibration_factors.get_index(sensor_id, to_dt)
     if from_dt_calibration_factors_index is None:
         raise ValueError(f"Calibration factors not found for {sensor_id} @ {from_dt}")
     if to_dt_calibration_factors_index is None:
@@ -129,7 +128,7 @@ def generate_geoms_file(
         os.remove(filepath)
 
     # apply calibration factors
-    cal = calibration_factors.root[to_dt_calibration_factors_index]
+    cal = geoms_metadata.calibration_factors.root[to_dt_calibration_factors_index]
     pl_df = pl_df.with_columns(
         pl.col("XCO2").mul(cal.xco2).alias("XCO2"),
         pl.col("XCH4").mul(cal.xch4).alias("XCH4"),
@@ -272,7 +271,6 @@ def generate_geoms_file(
 def run(
     config: Optional[src.types.Config] = None,
     geoms_metadata: Optional[src.types.GEOMSMetadata] = None,
-    calibration_factors: Optional[src.types.CalibrationFactorsList] = None,
 ) -> None:
     if config is None:
         print("Loading configuration")
@@ -283,104 +281,83 @@ def run(
         print("Loading geoms metadata")
         geoms_metadata = src.types.GEOMSMetadata.load()
 
-    if calibration_factors is None:
-        print("Loading calibration factors")
-        calibration_factors = src.types.CalibrationFactorsList.load()
-
-    for retrieval_algorithm in config.geoms.retrieval_algorithms:
-        for atmospheric_profile_model in config.geoms.atmospheric_profile_models:
-            if (retrieval_algorithm == "proffast-1.0") and (
-                atmospheric_profile_model == "GGG2020"
-            ):  # pragma: no cover
-                print("Skipping proffast-1.0/GGG2020 as it is not supported")
-                continue
-
-            print(f"Processing {retrieval_algorithm}/{atmospheric_profile_model}")
-
-            for _, sensor_id in enumerate(config.geoms.sensor_ids):
-                print(f'Processing sensor id "{sensor_id}"')
-
-                results_folders = os.path.join(
-                    config.general.data.results.root,
-                    retrieval_algorithm,
-                    atmospheric_profile_model,
-                    sensor_id,
-                    "successful",
-                )
-                if not os.path.exists(results_folders):  # pragma: no cover
-                    print(f"Sensor {sensor_id}: no results found")
+    for geoms_config in config.geoms:
+        for retrieval_algorithm in geoms_config.retrieval_algorithms:
+            for atmospheric_profile_model in geoms_config.atmospheric_profile_models:
+                if (retrieval_algorithm == "proffast-1.0") and (
+                    atmospheric_profile_model == "GGG2020"
[…]
