### 2026-09-10 b70215e
#154 Migrate configuration from JSON to TOML (10)
Refactor geoms and bundle config a bit more
