### 2026-09-10 265f1c6
#155 Integrate metadata library into this repository and also migrate to TOML (1)
Add current metadata library code to this repository
