+++ src/em27_metadata/types.py
@@ -1,0 +1,253 @@
+from __future__ import annotations
+from typing import Any, Optional
+import datetime
+import re
+import pydantic
+
+
+class TimeSeriesElement(pydantic.BaseModel):
+    model_config = pydantic.ConfigDict(arbitrary_types_allowed=True)
+
+    from_datetime: datetime.datetime = pydantic.Field(
+        ..., validation_alias=pydantic.AliasChoices("from_datetime", "from_dt")
+    )
+    to_datetime: datetime.datetime = pydantic.Field(
+        ..., validation_alias=pydantic.AliasChoices("to_datetime", "to_dt")
+    )
+
+    @pydantic.field_validator("from_datetime", "to_datetime", mode="before")
+    def datetime_string_validator(cls, v: str | datetime.datetime) -> datetime.datetime:
+        if isinstance(v, datetime.datetime):
+            return v
+        assert isinstance(v, str), "must be a string"
+        assert TimeSeriesElement.matches_datetime_regex(v), (
+            "must match the pattern YYYY-MM-DDTHH:MM:SS+HHMM"
+        )
+        return datetime.datetime.strptime(v, "%Y-%m-%dT%H:%M:%S%z")
+
+    @staticmethod
+    def matches_datetime_regex(v: str) -> bool:
+        return (
+            re.match(r"^(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2}):(\d{2})([\+\-])(\d{4})$", v)
+            is not None
+        )
+
+    @pydantic.model_validator(mode="after")
+    def model_validator(self) -> TimeSeriesElement:
+        if self.from_datetime > self.to_datetime:
+            raise ValueError(
+                f"from_datetime ({self.from_datetime}) > to_datetime ({self.to_datetime})"
+            )
+        if self.from_datetime.second != 0:
+            raise ValueError("from_datetime must be at the beginning of a minute (second=0)")
+        if self.to_datetime.second != 59:
+            raise ValueError("to_datetime must be at the end of a minute (second=59)")
+        return self
+
+    @pydantic.field_serializer("from_datetime", "to_datetime")
+    def t_serializer(self, dt: datetime.date, _info: Any) -> str:
+        return dt.strftime("%Y-%m-%dT%H:%M:%S+00:00")
+
+
+class Setup(pydantic.BaseModel):
+    location_id: str = pydantic.Field(
+        ...,
+        min_length=1,
+        description="Location ID referring to a location named in `locations.json`",
+        validation_alias=pydantic.AliasChoices("location_id", "lid"),
+    )
+    pressure_data_source: Optional[str] = pydantic.Field(
+        default=None,
+        min_length=1,
+        description="Pressure data source, if not set, using the pressure of the sensor",
+        validation_alias=pydantic.AliasChoices("pressure_data_source", "pds"),
+    )
+    utc_offset: float = pydantic.Field(
+        default=0,
+        gt=-12,
+        lt=12,
+        description="UTC offset of the location, if not set, using an offset of 0",
+    )
+    atmospheric_profile_location_id: Optional[str] = pydantic.Field(
+        default=None,
+        min_length=1,
+        description="Location ID referring to a location named in `locations.json`. This location's coordinates are used for the atmospheric profiles in the retrieval.",
+        validation_alias=pydantic.AliasChoices("atmospheric_profile_location_id", "profile_lid"),
+    )
+
+
+class SetupsListItem(TimeSeriesElement):
+    """An element in the `sensor.setups` list"""
+
+    value: Setup = pydantic.Field(..., validation_alias=pydantic.AliasChoices("value", "v"))
+
+
+class LocationMetadata(pydantic.BaseModel):
+    location_id: str = pydantic.Field(
+        ...,
+        min_length=1,
[…]
