### 2026-09-10 4848495
#154 Migrate configuration from JSON to TOML (11)
Rename subproject directories: "geoms" -> "geom_exporter", and so on
