### 2026-09-10 552ccae
#155 Integrate metadata library into this repository and also migrate to TOML (2)
Change all import paths to the repo library and do remove the dependency
