+++ src/retrieval/session/create_session.py
@@ -1,10 +1,8 @@
 import datetime
 import os
-
-import em27_metadata
 import tum_esm_utils
 
-from src import retrieval, types
+from src import retrieval, types, em27_metadata
 
 _PROJECT_DIR = tum_esm_utils.files.get_parent_dir_path(__file__, current_depth=4)
 _RETRIEVAL_ALGORITHMS_DIR = os.path.join(_PROJECT_DIR, "src", "retrieval", "algorithms")
