+++ src/__init__.py
@@ -1,4 +1,5 @@
 from . import bundle_exporter as bundle_exporter
+from . import em27_metadata as em27_metadata
 from . import geoms_exporter as geoms_exporter
 from . import ggg_profiles_downloader as ggg_profiles_downloader
 from . import retrieval as retrieval
