### 2026-09-10 8a04f20
#155 Integrate metadata library into this repository and also migrate to TOML (3)
Add support for loading both old and new metadata schema from github.

Use github file loader from tum-esm-utils library
