### 2026-09-10 389aa40
#155 Integrate metadata library into this repository and also migrate to TOML (4)
Refine metadata interface -> campaigns should be optional everywhere
