### 2026-09-10 0de02e4
#155 Integrate metadata library into this repository and also migrate to TOML (5)
Make local file loader support both old and new format
