### 2026-09-10 70b8812
#155 Integrate metadata library into this repository and also migrate to TOML (6)
Make example data loader load sample data in new formats and also test fallback
