+++ src/em27_metadata/sample_data/README.md
@@ -1,3 +1,3 @@
 DO NOT REMOVE THIS DIRECTORY.
 
-It is a duplicate of the `data` directory, but it is required in this location because it is shipped with the library bundle.
+It is required in this location because it is shipped with the library bundle.
