+++ src/em27_metadata/sample_data/toml/em27_metadata.toml
@@ -1,0 +1,59 @@
+# LOCATIONS
+
+[[locations]]
+location_id = "lid1"
+details = "description of location 1"
+lon = 10.5
+lat = 48.1
+alt = 500
+
+[[locations]]
+location_id = "lid2"
+details = "description of location 2"
+lon = 11.3
+lat = 48.0
+alt = 600
+
+# SENSORS
+
+[[sensors]]
+sensor_id = "sid1"
+serial_number = 50
+setups = [
+    { from_datetime = "2020-08-22T00:00:00+0000", to_datetime = "2020-08-25T23:59:59+0000", value = { location_id = "lid1", utc_offset = 2, pressure_data_source = "LMU-MIM01-height-adjusted", atmospheric_profile_location_id = "lid2" } },
+    { from_datetime = "2020-08-26T00:00:00+0000", to_datetime = "2020-08-30T23:59:59+0000", value = { location_id = "lid1", utc_offset = 2, pressure_data_source = "LMU-MIM01-height-adjusted" } },
+    { from_datetime = "2020-08-31T00:00:00+0000", to_datetime = "2020-09-26T23:59:59+0000", value = { location_id = "lid1", utc_offset = 2 } },
+    { from_datetime = "2020-09-27T00:00:00+0000", to_datetime = "2020-10-01T23:59:59+0000", value = { location_id = "lid2" } }
+]
+
+[[sensors]]
+sensor_id = "sid2"
+serial_number = 51
+setups = [
+    { from_datetime = "2020-08-26T00:00:00+0000", to_datetime = "2020-10-01T23:59:59+000", value = { location_id = "lid1" } }
+]
+
+# CAMPAIGNS
+
+[[campaigns]]
+campaign_id = "cid1"
+from_datetime = "2019-09-13T00:00:00+0000"
+to_datetime = "2100-01-01T23:59:59+0000"
+sensor_ids = ["sid1", "sid2"]
+location_ids = ["lid1", "lid2"]
+
+# EVENTS
+
+[[events]]
+sensor_ids = ["sid1", "sid2"]
+from_datetime = "2024-06-24T10:30:00+0000"
+to_datetime = "2024-06-24T11:29:59+0000"
+description = "Mirror Change - Sensors have to be recalibrated"
+data_is_usable = false
+
+[[events]]
+sensor_ids = ["sid1"]
+from_datetime = "2025-08-24T10:30:00+0000"
+to_datetime = "2025-08-24T17:29:59+0000"
+description = "ILS Measurements"
+data_is_usable = false
