### 2026-09-10 f286153
#154 Migrate configuration from JSON to TOML (12)
Refactor all tests after the schema and repository refactor
