+++ tests/ggg_profiles_downloader/test_list_requested_data.py
@@ -1,5 +1,4 @@
 import datetime
-import em27_metadata
 import pytest
 import tum_esm_utils
 import src
@@ -9,62 +8,62 @@ from ..fixtures import provide_config_template  # pyright: ignore[reportUnusedIm
 @pytest.mark.order(3)
 @pytest.mark.quick
 def test_list_requested_data(provide_config_template: src.types.Config) -> None:
-    metadata = em27_metadata.EM27MetadataInterface(
-        locations=em27_metadata.types.LocationMetadataList(
+    metadata = src.em27_metadata.EM27MetadataInterface(
+        locations=src.em27_metadata.types.LocationMetadataList(
             root=[
-                em27_metadata.types.LocationMetadata(
+                src.em27_metadata.types.LocationMetadata(
                     location_id="l1", details="l1 details", lat=1, lon=2, alt=0
                 ),
-                em27_metadata.types.LocationMetadata(
+                src.em27_metadata.types.LocationMetadata(
                     location_id="l2", details="l2 details", lat=1, lon=3, alt=0
                 ),
-                em27_metadata.types.LocationMetadata(
+                src.em27_metadata.types.LocationMetadata(
                     location_id="l3", details="l3 details", lat=2, lon=3, alt=0
                 ),
             ]
         ),
-        sensors=em27_metadata.types.SensorMetadataList(
+        sensors=src.em27_metadata.types.SensorMetadataList(
             root=[
-                em27_metadata.types.SensorMetadata(
+                src.em27_metadata.types.SensorMetadata(
                     sensor_id="s1",
                     serial_number=1,
                     setups=[
-                        em27_metadata.types.SetupsListItem(
+                        src.em27_metadata.types.SetupsListItem(
                             from_datetime="2000-01-01T00:00:00+0000",  # pyright: ignore[reportArgumentType]
                             to_datetime="2000-03-01T11:59:59+0000",  # pyright: ignore[reportArgumentType]
-                            value=em27_metadata.types.Setup(location_id="l1"),
+                            value=src.em27_metadata.types.Setup(location_id="l1"),
                         ),
-                        em27_metadata.types.SetupsListItem(
+                        src.em27_metadata.types.SetupsListItem(
                             from_datetime="2000-03-01T12:00:00+0000",  # pyright: ignore[reportArgumentType]
                             to_datetime="2000-05-01T23:59:59+0000",  # pyright: ignore[reportArgumentType]
-                            value=em27_metadata.types.Setup(location_id="l3"),
+                            value=src.em27_metadata.types.Setup(location_id="l3"),
                         ),
-                        em27_metadata.types.SetupsListItem(
+                        src.em27_metadata.types.SetupsListItem(
                             from_datetime="2000-05-04T12:00:00+0000",  # pyright: ignore[reportArgumentType]
                             to_datetime="2000-05-07T23:59:59+0000",  # pyright: ignore[reportArgumentType]
-                            value=em27_metadata.types.Setup(location_id="l2"),
+                            value=src.em27_metadata.types.Setup(location_id="l2"),
                         ),
                     ],
                 ),
-                em27_metadata.types.SensorMetadata(
+                src.em27_metadata.types.SensorMetadata(
                     sensor_id="s2",
                     serial_number=2,
                     setups=[
-                        em27_metadata.types.SetupsListItem(
+                        src.em27_metadata.types.SetupsListItem(
                             from_datetime="2000-01-07T00:00:00+0000",  # pyright: ignore[reportArgumentType]
                             to_datetime="2000-02-23T23:59:59+0000",  # pyright: ignore[reportArgumentType]
[…]
