### 2026-09-10 28b665c
#155 Integrate metadata library into this repository and also migrate to TOML (7)
Add all tests from old repo in here
