+++ tests/em27_metadata/test_datetime_parsing.py
@@ -1,0 +1,41 @@
+import pytest
+import src
+
+
+@pytest.mark.em27_metadata_library
+def test_datetime_parsing() -> None:
+    assert src.em27_metadata.types.TimeSeriesElement.matches_datetime_regex(
+        "2016-10-01T00:00:00+0000"
+    )
+    assert src.em27_metadata.types.TimeSeriesElement.matches_datetime_regex(
+        "2016-10-01T23:00:00+0000"
+    )
+    assert src.em27_metadata.types.TimeSeriesElement.matches_datetime_regex(
+        "2016-10-01T00:00:00+0100"
+    )
+    assert not src.em27_metadata.types.TimeSeriesElement.matches_datetime_regex(
+        "2016-10-01T00:00:00+000"
+    )
+
+
+@pytest.mark.em27_metadata_library
+def test_time_series_element() -> None:
+    tse1 = src.em27_metadata.types.TimeSeriesElement(
+        from_datetime="2016-10-01T00:00:00+0000",  # pyright: ignore[reportArgumentType]
+        to_datetime="2016-10-01T22:59:59+0200",  # pyright: ignore[reportArgumentType]
+    )
+    actual_dt_seconds = (tse1.to_datetime - tse1.from_datetime).total_seconds()
+    expected_dt_seconds = 3600 * 21 - 1
+    assert actual_dt_seconds == expected_dt_seconds, (
+        f"dt_seconds: {actual_dt_seconds} (actual) != {expected_dt_seconds} (expected)"
+    )
+
+    tse2 = src.em27_metadata.types.TimeSeriesElement(
+        from_datetime="2016-10-01T00:00:00+0000",  # pyright: ignore[reportArgumentType]
+        to_datetime="2016-10-03T13:24:59-0530",  # pyright: ignore[reportArgumentType]
+    )
+    actual_dt_seconds = (tse2.to_datetime - tse2.from_datetime).total_seconds()
+    expected_dt_seconds = 61 * 3600 + (24 * 60) + 59 + (5 * 3600) + (30 * 60)
+    assert actual_dt_seconds == expected_dt_seconds, (
+        f"dt_seconds: {actual_dt_seconds} (actual) != {expected_dt_seconds} (expected)"
+    )
