+++ pyproject.toml
@@ -75,6 +75,7 @@ markers = [
     "quick: can be run in a CI environment, < 30 seconds, everything except running actual retrievals",
     "ci: can be run in a CI environment, ~ 1 - 2 min, runs the complete tests but without running the actual retrieval algorithms",
     "complete: can be run in a CI environment, ~ 3 - 10 minutes, runs three sample days from different instruments and locations for every retrieval algorithm and every atmospheric profile",
+    "em27_metadata_library: is testing the em27_metadata library, not really required to be run by users"
 ]
 addopts = "--exitfirst --verbose --durations=0 --no-header --cov=src"
 

[…]
