### 2026-09-10 918c5dc
#155 Integrate metadata library into this repository and also migrate to TOML (8)
Add metadata template and test the template
