### 2026-09-10 5ae710f
#155 Integrate metadata library into this repository and also migrate to TOML (9)
Debug library tests
