+++ tests/em27_metadata/test_loaders.py
@@ -1,9 +1,10 @@
 import os
 import pytest
+import tum_esm_utils
 import src
 import dotenv
 
-PROJECT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+PROJECT_DIR = tum_esm_utils.files.rel_to_abs_path("../..")
 
 
 @pytest.mark.em27_metadata_library
@@ -17,15 +18,15 @@ def test_remote_loader_private_repo() -> None:
     if os.path.exists(env_file):
         dotenv.load_dotenv(env_file)
     access_token = os.getenv("TUM_ESM_GITHUB_ACCESS_TOKEN")
-    if access_token is None:
-        pytest.skip("TUM_ESM_GITHUB_ACCESS_TOKEN not set as environment variable")
+    # if access_token is None:
+    #    pytest.skip("TUM_ESM_GITHUB_ACCESS_TOKEN not set as environment variable")
+    assert access_token is not None
     src.em27_metadata.load_from_github("tum-esm/em27-metadata-storage", access_token=access_token)
 
 
 @pytest.mark.em27_metadata_library
 def test_local_loader() -> None:
-    src.em27_metadata.load_from_local_files(
-        locations_path=os.path.join(PROJECT_DIR, "data", "locations.json"),
-        sensors_path=os.path.join(PROJECT_DIR, "data", "sensors.json"),
-        campaigns_path=os.path.join(PROJECT_DIR, "data", "campaigns.json"),
-    )
+    for v in ["toml", "json"]:
+        src.em27_metadata.load_from_local_files(
+            config_directory=os.path.join(PROJECT_DIR, "src", "em27_metadata", "sample_data", v),
+        )
