+++ src/em27_metadata/sample_data/toml/em27_metadata.toml
@@ -30,7 +30,7 @@ setups = [
 sensor_id = "sid2"
 serial_number = 51
 setups = [
-    { from_datetime = "2020-08-26T00:00:00+0000", to_datetime = "2020-10-01T23:59:59+000", value = { location_id = "lid1" } }
+    { from_datetime = "2020-08-26T00:00:00+0000", to_datetime = "2020-10-01T23:59:59+0000", value = { location_id = "lid1" } }
 ]
 
 # CAMPAIGNS
