### 2026-09-10 6048b01
#154 Migrate configuration from JSON to TOML (13)
Add support to automatically load old config files to new TOML schema
