+++ src/types/old_schemas.py
@@ -11,7 +11,7 @@ from .basic_types import AtmosphericProfileModel, RetrievalAlgorithm
 # CONFIG
 
 
-class MetadataConfig(pydantic.BaseModel):
+class _MetadataConfig(pydantic.BaseModel):
     """GitHub repository where the location data is stored."""
 
     model_config = pydantic.ConfigDict(extra="forbid")
@@ -28,7 +28,7 @@ class MetadataConfig(pydantic.BaseModel):
     )
 
 
-class GroundPressureConfig(pydantic.BaseModel):
+class _GroundPressureConfig(pydantic.BaseModel):
     """Format of the ground pressure files. We support any text file that stores one data point per row and separates the columns with a comma, space, or tab, i.e. CSV, TSV, or space-separated files. Using the `file_regex` field, you specify which files to consider for a given sensor id and date.
 
     You have to specify the columns that contain the date and time of the data. There is three options to specify this - the CLI will complain if you configure none or more than one of these options:
@@ -135,7 +135,7 @@ class GroundPressureConfig(pydantic.BaseModel):
 
     # validate -> you can only set either date AND time OR unix_timestamp
     @pydantic.model_validator(mode="after")
-    def _check_datetime_columns(self) -> GroundPressureConfig:
+    def _check_datetime_columns(self) -> _GroundPressureConfig:
         required = {
             "datetime_column": [],
             "date_column": ["time_column"],
@@ -167,12 +167,12 @@ class GroundPressureConfig(pydantic.BaseModel):
         return self
 
 
-class DataConfig(pydantic.BaseModel):
+class _DataConfig(pydantic.BaseModel):
     """Location where the input data sourced from."""
 
     model_config = pydantic.ConfigDict(extra="forbid")
 
-    ground_pressure: GroundPressureConfig = pydantic.Field(
+    ground_pressure: _GroundPressureConfig = pydantic.Field(
         ...,
         description="directory path and format configuration of the ground pressure files",
     )
@@ -197,7 +197,7 @@ class DataConfig(pydantic.BaseModel):
         return values
 
 
-class ProfilesServerConfig(pydantic.BaseModel):
+class _ProfilesServerConfig(pydantic.BaseModel):
     """Settings for accessing the ccycle ftp server. Besides the
     `email` field, these can be left as default in most cases."""
 
@@ -216,7 +216,7 @@ class ProfilesServerConfig(pydantic.BaseModel):
     )
 
 
-class ProfilesScopeConfig(pydantic.BaseModel):
+class _ProfilesScopeConfig(pydantic.BaseModel):
     model_config = pydantic.ConfigDict(extra="forbid")
 
     from_date: datetime.date = pydantic.Field(
@@ -237,13 +237,13 @@ class ProfilesScopeConfig(pydantic.BaseModel):
     )
 
     @pydantic.model_validator(mode="after")
-    def check_date_order(self) -> ProfilesScopeConfig:
+    def check_date_order(self) -> _ProfilesScopeConfig:
         if self.from_date > self.to_date:
             raise ValueError("from_date must be before to_date")
         return self
 
 
-class ProfilesGGG2020StandardSitesItemConfig(pydantic.BaseModel):
+class _ProfilesGGG2020StandardSitesItemConfig(pydantic.BaseModel):
     model_config = pydantic.ConfigDict(extra="forbid")
 
     identifier: str = pydantic.Field(
@@ -270,13 +270,13 @@ class ProfilesGGG2020StandardSitesItemConfig(pydantic.BaseModel):
     )
 
     @pydantic.model_validator(mode="after")
-    def check_date_order(self) -> ProfilesGGG2020StandardSitesItemConfig:
[…]
