### 2026-09-10 acfb90e
#154 Migrate configuration from JSON to TOML (14)
Add support to load new geoms schema from old geoms schema
