+++ src/types/old_schemas.py
@@ -613,7 +613,7 @@ class OldConfig(pydantic.BaseModel):
 # GEOMS METADATA
 
 
-class GEOMSGeneralMetadata(pydantic.BaseModel):
+class _GEOMSGeneralMetadata(pydantic.BaseModel):
     network: str = pydantic.Field(
         ...,
         description="Used in the filename of the HDF5 file",
@@ -631,7 +631,7 @@ class GEOMSGeneralMetadata(pydantic.BaseModel):
     )
 
 
-class GEOMSDataMetadata(pydantic.BaseModel):
+class _GEOMSDataMetadata(pydantic.BaseModel):
     discipline: str = pydantic.Field(
         ...,
         description="The value of the HDF5 attribute `DATA_DISCIPLINE`",
@@ -659,7 +659,7 @@ class GEOMSDataMetadata(pydantic.BaseModel):
     )
 
 
-class GEOMSFileMetadata(pydantic.BaseModel):
+class _GEOMSFileMetadata(pydantic.BaseModel):
     doi: str = pydantic.Field(
         ...,
         description="The value of the HDF5 attribute `FILE_DOI`",
@@ -681,7 +681,7 @@ class GEOMSFileMetadata(pydantic.BaseModel):
     )
 
 
-class GEOMSContactlMetadata(pydantic.BaseModel):
+class _GEOMSContactlMetadata(pydantic.BaseModel):
     name: str = pydantic.Field(
         ...,
         description="The value of the HDF5 attribute `PI_NAME`/`DO_NAME`/`DS_NAME`",
@@ -706,37 +706,22 @@ class GEOMSContactlMetadata(pydantic.BaseModel):
     )
 
 
-class GEOMSMetadata(pydantic.BaseModel):
-    general: GEOMSGeneralMetadata
-    data: GEOMSDataMetadata
-    file: GEOMSFileMetadata
-    principle_investigator: GEOMSContactlMetadata
-    data_originator: GEOMSContactlMetadata
-    data_submitter: GEOMSContactlMetadata
+class OldGEOMSMetadata(pydantic.BaseModel):
+    general: _GEOMSGeneralMetadata
+    data: _GEOMSDataMetadata
+    file: _GEOMSFileMetadata
+    principle_investigator: _GEOMSContactlMetadata
+    data_originator: _GEOMSContactlMetadata
+    data_submitter: _GEOMSContactlMetadata
     locations: dict[str, str] = pydantic.Field(
         ..., description="Maps your locations id to the corresponding EVDC location id"
     )
-
-    @staticmethod
-    def load(template: bool = False) -> GEOMSMetadata:
-        """Load the EVDC metadata from `<config_dir>/geoms_metadata.json`."""
-
-        erp_config_dir = tum_esm_utils.files.rel_to_abs_path("../../config")
-        if not template:
-            env_path = os.path.join(tum_esm_utils.files.rel_to_abs_path("../../config"), ".env")
-            if os.path.isfile(env_path):
-                dotenv.load_dotenv(env_path)
-            erp_config_dir = os.getenv("ERP_CONFIG_DIR", erp_config_dir)
-        filepath = os.path.join(
-            erp_config_dir, f"geoms_metadata{'.template' if template else ''}.json"
-        )
-        return GEOMSMetadata.model_validate_json(tum_esm_utils.files.load_file(filepath))
 
 
 # CALIBRATION FACTORS
 
 
-class CalibrationFactors(pydantic.BaseModel):
+class _CalibrationFactors(pydantic.BaseModel):
     model_config = pydantic.ConfigDict(arbitrary_types_allowed=True)
 
     sensor_id: str
[…]
