### 2026-09-10 ace0d49
#154 Migrate configuration from JSON to TOML (15)
Debug some things
