### 2026-09-10 71edada
#154 Migrate configuration from JSON to TOML (17)
Debug
