+++ src/utils/metadata.py
@@ -6,8 +6,16 @@ from src import types, em27_metadata
 def load_local_em27_metadata_interface(
     config_dir: str = types.Config.get_config_dir(),
 ) -> Optional[em27_metadata.EM27MetadataInterface]:
+    config_dir = os.path.abspath(config_dir)
     if not os.path.isdir(config_dir):
         return None
+
+    # TODO: refactor this loading
+
+    toml_path = os.path.join(config_dir, "em27_metadata.toml")
+    if os.path.isfile(toml_path):
+        return em27_metadata.loader.load_from_local_files(config_directory=config_dir)
+
     locations_path = os.path.join(config_dir, "locations.json")
     sensors_path = os.path.join(config_dir, "sensors.json")
     campaigns_path = os.path.join(config_dir, "campaigns.json")
@@ -23,11 +31,6 @@ def load_local_em27_metadata_interface(
                 + "tions.json, sensors.json). Please "
                 + "add or remove all local metadata files."
             )
-        return em27_metadata.loader.load_from_local_files(
-            locations_path=locations_path,
-            sensors_path=sensors_path,
-            campaigns_path=campaigns_path if os.path.isfile(campaigns_path) else None,
-            events_path=events_path if os.path.isfile(events_path) else None,
-        )
+        return em27_metadata.loader.load_from_local_files(config_directory=config_dir)
     else:
         return None
