+++ config/em27_metadata.template.toml
@@ -20,14 +20,14 @@ alt_asl = 180.0
 sensor_id = "so"
 serial_number = 39
 setups = [
-    { from_datetime = "2017-06-08T00:00:00+0000", to_datetime = "2017-06-09T23:59:59+0000", location_id = "SOD", utc_offset = 0 }
+    { from_datetime = "2017-06-08T00:00:00+0000", to_datetime = "2017-06-09T23:59:59+0000", value = { location_id = "SOD", utc_offset = 0 } }
 ]
 
 [[sensors]]
 sensor_id = "mc"
 serial_number = 115
 setups = [
-    { from_datetime = "2022-06-02T00:00:00+0000", to_datetime = "2022-06-02T23:59:59+0000", location_id = "ZEN", utc_offset = 0 }
+    { from_datetime = "2022-06-02T00:00:00+0000", to_datetime = "2022-06-02T23:59:59+0000", value = { location_id = "ZEN", utc_offset = 0 } }
 ]
 
 # CAMPAIGNS

[…]
