### 2026-09-10 a0b8b65
#154 Migrate configuration from JSON to TOML (18)
Automatically test config and geoms_metadata conversion
