### 2026-09-10 55a399b
#155 Integrate metadata library into this repository and also migrate to TOML (10)
Automatically test em27 metadata conversion
