### 2026-09-10 904661c
#154 Migrate configuration from JSON to TOML (19)
Debug
