+++ config/old_templates/config.automatically_converted.toml
@@ -72,7 +72,6 @@ pressure_calibration_offsets = { mb = -0.000125 }
 
 [retrieval.jobs.0.custom_ils]
 ma = { channel1_me = 0.9892, channel1_pe = -0.001082, channel2_me = 0.9892, channel2_pe = -0.001082 }
-mb = { channel1_me = 0.9893, channel1_pe = -0.001083, channel2_me = 0.9893, channel2_pe = -0.001083 }
 
 [retrieval.jobs.1]
 retrieval_algorithm = "proffast-2.3"
