### 2026-09-10 d6f9b63
#155 Integrate metadata library into this repository and also migrate to TOML (11)
Fix typing issues
