### 2026-09-11 1ae194d
Refine docs (2)
Host docs on GitHub pages instead of netlify
