+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,5 +1,5 @@
 blank_issues_enabled: true
 contact_links:
   - name: Documentation
-    url: https://em27-retrieval-pipeline.netlify.app/
+    url: https://tum-esm.github.io/em27-retrieval-pipeline/
     about: Please read the documentation before submitting an issue.
