### 2026-09-11 f4e11f6
Refine docs (4)
Update sync script and sync new schemas and templates
