+++ docs/sync.py
@@ -1,0 +1,259 @@
+import os
+import re
+import shutil
+import sys
+import copy
+from typing import Any, Optional
+import click
+import tum_esm_utils.files
+
+sys.path.append(tum_esm_utils.files.rel_to_abs_path(".."))
+
+import src
+import cli
+
+PROJECT_DIR = tum_esm_utils.files.rel_to_abs_path("..")
+
+
+def simplify_anyof_nullable(schema: Any) -> Any:
+    """
+    Recursively simplify JSON Schema fragments like:
+
+        {"anyOf": [{"type": "string", ...}, {"type": "null"}], ...}
+
+    into:
+
+        {"type": "string", "nullable": true, ...}
+
+    Rules:
+    - Only simplifies `anyOf` if there is exactly one non-null branch.
+    - Copies inner keys from the non-null branch to the outer object.
+    - Outer keys win if there is a conflict.
+    - Outer `description` is never overwritten.
+    - Adds `"nullable": true` if one of the branches is `{"type": "null"}`.
+    """
+    if isinstance(schema, list):
+        return [simplify_anyof_nullable(item) for item in schema]
+
+    if not isinstance(schema, dict):
+        return schema
+
+    # First recurse into children
+    schema = {k: simplify_anyof_nullable(v) for k, v in schema.items()}
+
+    anyof = schema.get("anyOf")
+    if isinstance(anyof, list):
+        non_null_options = []
+        has_null = False
+
+        for option in anyof:
+            if isinstance(option, dict) and option.get("type") == "null":
+                has_null = True
+            else:
+                non_null_options.append(option)
+
+        if len(non_null_options) == 1 and isinstance(non_null_options[0], dict):
+            inner = copy.deepcopy(non_null_options[0])
+
+            # Remove anyOf from outer before merging
+            schema.pop("anyOf", None)
+
+            # Merge inner keys into outer, but do not overwrite outer keys
+            for key, value in inner.items():
+                if key == "description" and "description" in schema:
+                    continue
+                if key not in schema:
+                    schema[key] = value
+
+            if has_null:
+                schema["nullable"] = True
+
+    return schema
+
+
+def simplify_enum_nullable(schema: Any) -> Any:
+    """
+    Recursively simplify JSON Schema fragments like:
+
+        {"enum": ["a", "b", null], ...}
+
+    into:
+
+        {"enum": ["a", "b", null], "type": "string", "nullable": true, ...}
+
+    Rules:
+    - Only simplifies `enum` if values contain `null` and all non-null values
+      are of exactly one JSON type.
+    - Adds `"type"` only if it does not already exist.
+    - Adds `"nullable": true` when `null` is part of the enum.
+    """
+    if isinstance(schema, list):
+        return [simplify_enum_nullable(item) for item in schema]
+
+    if not isinstance(schema, dict):
+        return schema
+
+    # First recurse into children
+    schema = {k: simplify_enum_nullable(v) for k, v in schema.items()}
+
+    enum_values = schema.get("enum")
+    if not isinstance(enum_values, list):
+        return schema
+
+    has_null = any(v is None for v in enum_values)
+    if not has_null:
+        return schema
+
+    non_null_values = [v for v in enum_values if v is not None]
+    if not non_null_values:
+        return schema
+
+    def json_type(value: Any) -> str:
+        # bool is a subclass of int in Python, so detect it first.
+        if isinstance(value, bool):
+            return "boolean"
+        if isinstance(value, int):
+            return "integer"
+        if isinstance(value, float):
[…]
