### 2026-09-11 b85dea0
Refine docs (5)
Move "api-reference" docs to "reference" path

Remove old syncing script
