+++ src/scripts/sync_api_reference.py
@@ -1,151 +1,0 @@
-import os
-import re
-import shutil
-import sys
-from typing import Optional
-
-import click.core
-import tum_esm_utils
-
-_PROJECT_DIR = tum_esm_utils.files.rel_to_abs_path("../..")
-_DOCS_DIR = os.path.join(_PROJECT_DIR, "docs")
-
-sys.path.append(_PROJECT_DIR)
-import cli
-import src
-
-# ---------------------------------------------------------
-# EXPORT CLI REFERENCE
-
-# credits to https://stackoverflow.com/a/58018765/8255842
-
-
-def recursive_help(
-    command: click.Group | click.Command,
-    parent_context: Optional[click.core.Context] = None,
-) -> str:
-    output: str = ""
-    context = click.core.Context(command, info_name=command.name, parent=parent_context)
-    if isinstance(command, click.Group):
-        for sub_command in command.commands.values():
-            output += recursive_help(sub_command, parent_context=context)
-    else:
-        assert command.short_help is not None, (
-            f"Command {context.command_path} has no short help".format(command=command)
-        )
-        output += f"## {command.short_help}\n\n"
-        help_text = command.get_help(context)
-        # find all options (--help)
-        options: list[str] = re.findall(r"(\-\-\w[\w\-]+ )", help_text)
-        for option in options:
-            help_text = help_text.replace(option, f"\n`{option}`")
-        output += (
-            help_text.replace(
-                "\n  ",
-                "\n",
-            )
-            .replace(
-                "[OPTIONS]",
-                "",
-            )
-            .replace(
-                f"Usage: {context.command_path}",
-                f"**Usage:**\n\n`python cli.py {context.command_path[4:]} [OPTIONS]`\n\n**Description:** ",
-            )
-            .replace(
-                "Options:\n",
-                "**Options:**\n\n",
-            )
-            + "\n\n"
-        )
-    return output
-
-
-print("Exporting CLI reference to docs/src/content/api-reference/cli.mdx")
-with open(f"{_DOCS_DIR}/src/content/docs/api-reference/cli.mdx", "w") as f:
-    f.write("---\ntitle: CLI Reference\n---\n\n")
-    f.write(recursive_help(cli.cli))
-
-# ---------------------------------------------------------
-# EXPORT JSON SCHEMAS
-
-print("Exporting JSON schemas to docs/src/assets")
-tum_esm_utils.files.dump_json_file(
-    f"{_DOCS_DIR}/src/assets/config.schema.json",
-    src.types.Config.model_json_schema(mode="validation"),
-)
-tum_esm_utils.files.dump_json_file(
-    f"{_DOCS_DIR}/src/assets/locations.schema.json",
-    src.em27_metadata.types.LocationMetadataList.model_json_schema(mode="validation"),
-)
-tum_esm_utils.files.dump_json_file(
-    f"{_DOCS_DIR}/src/assets/sensors.schema.json",
-    src.em27_metadata.types.SensorMetadataList.model_json_schema(mode="validation"),
-)
-tum_esm_utils.files.dump_json_file(
-    f"{_DOCS_DIR}/src/assets/campaigns.schema.json",
-    src.em27_metadata.types.CampaignMetadataList.model_json_schema(mode="validation"),
-)
-tum_esm_utils.files.dump_json_file(
-    f"{_DOCS_DIR}/src/assets/events.schema.json",
-    src.em27_metadata.types.EventMetadataList.model_json_schema(mode="validation"),
-)
[…]
