### 2026-09-11 5e72bee
Refine docs (6)
Add the heavily refined API reference renderer
