### 2026-09-11 fafbfcb
Ensure TOML 1.1 support
