+++ pyproject.toml
@@ -13,7 +13,7 @@ dependencies = [
     "filelock>=3.16.1",
     "pydantic>=2.10.2",
     "tum-esm-utils[opus]>=2.11.3",
-    "tomli>=2.2.1",
+    "tomli>=2.4.0",
     "python-dotenv>=1.0.1",
     "skyfield>=1.49",              # used by the retrieval
     "rich>=13.9.4",                # used by export and profiles download
