### 2026-09-11 3e229a4
Refine docs (10)
Add the full config page
