### 2026-09-12 f9870be
Adopt datetime string logic from RETRO (1)
all configs contain datetimes as string with a clear documented regex pattern instead of relying on automated parsing of these strings in any allowed TOML format.
