+++ src/types/config.py
@@ -9,7 +9,14 @@ import pydantic
 import tomli
 import tum_esm_utils
 
-from .basic_types import AtmosphericProfileModel, RetrievalAlgorithm
+from .basic_types import (
+    DATE_STRING_PATTERN,
+    UTC_DATETIME_STRING_PATTERN,
+    AtmosphericProfileModel,
+    RetrievalAlgorithm,
+    parse_date_string,
+    parse_datetime_string,
+)
 from .old_schemas import OldConfig
 
 # TODO: refactor metadata source logic
@@ -271,12 +278,14 @@ class GGGProfilesDownloaderSubConfigs:
     class Scope(pydantic.BaseModel):
         model_config = pydantic.ConfigDict(extra="forbid")
 
-        from_date: datetime.date = pydantic.Field(
-            datetime.date(1900, 1, 1),
+        from_date: str = pydantic.Field(
+            "1900-01-01",
+            pattern=DATE_STRING_PATTERN,
             description="Date in format `YYYY-MM-DD` from which to request vertical profile data.",
         )
-        to_date: datetime.date = pydantic.Field(
-            datetime.date(2100, 1, 1),
+        to_date: str = pydantic.Field(
+            "2100-01-01",
+            pattern=DATE_STRING_PATTERN,
             description="Date in format `YYYY-MM-DD` until which to request vertical profile data.",
         )
         models: list[AtmosphericProfileModel] = pydantic.Field(
@@ -284,15 +293,23 @@ class GGGProfilesDownloaderSubConfigs:
             description="list of data types to request from the ccycle ftp server.",
         )
         force_download_locations: list[str] = pydantic.Field(
-            [],
+            default=[],
             description="List of locations to force download data for. These will be downloaded even at times where no instrument in the metadata is located there.",
         )
 
         @pydantic.model_validator(mode="after")
         def check_date_order(self) -> GGGProfilesDownloaderSubConfigs.Scope:
-            if self.from_date > self.to_date:
+            if self.from_date_parsed > self.to_date_parsed:
                 raise ValueError("from_date must be before to_date")
             return self
+
+        @property
+        def from_date_parsed(self) -> datetime.date:
+            return parse_date_string(self.from_date)
+
+        @property
+        def to_date_parsed(self) -> datetime.date:
+            return parse_date_string(self.to_date)
 
     class GGG2020StandardSitesItem(pydantic.BaseModel):
         model_config = pydantic.ConfigDict(extra="forbid")
@@ -312,20 +329,32 @@ class GGGProfilesDownloaderSubConfigs:
             ge=-180,
             le=180,
         )
-        from_date: datetime.date = pydantic.Field(
+        from_date: str = pydantic.Field(
             ...,
+            pattern=DATE_STRING_PATTERN,
             description="Date in format `YYYY-MM-DD` from which this standard site is active.",
         )
-        to_date: datetime.date = pydantic.Field(
-            default_factory=lambda: datetime.date.today() - datetime.timedelta(days=1),
+        to_date: str = pydantic.Field(
+            default_factory=lambda: (datetime.date.today() - datetime.timedelta(days=1)).strftime(
+                "%Y-%m-%d"
+            ),
+            pattern=DATE_STRING_PATTERN,
             description="Date in format `YYYY-MM-DD` until which this standard site is active. Default is yesterday.",
         )
 
         @pydantic.model_validator(mode="after")
         def check_date_order(self) -> GGGProfilesDownloaderSubConfigs.GGG2020StandardSitesItem:
-            if self.from_date > self.to_date:
+            if self.from_date_parsed > self.to_date_parsed:
                 raise ValueError("from_date must be before to_date")
             return self
+
+        @property
[…]
