### 2026-09-12 725e32f
Adopt datetime string logic from RETRO (2)
add tests for this and update schema documentation
