### 2026-09-12 3d1b8e7
Make all quick tests run again
