### 2026-09-12 e86974d
#155 Integrate metadata library into this repository and also migrate to TOML (12)
Use new TOML-based metadata in example
