### 2026-09-12 e167aa6
#155 Integrate metadata library into this repository and also migrate to TOML (13)
Clean up the sensor metadata schema.

Now "sensor.setups" is called "sensor.deployments" and the values nested in "sensor.deployments[i].value are now directly stored in the ""sensor.deployments[i]" object.
