+++ tests/em27_metadata/test_explosion.py
@@ -24,41 +24,45 @@ def test_explosion() -> None:
             ),
         ]
     )
-    # fmt: off
     sensors = src.em27_metadata.types.SensorMetadataList(
         root=[
             src.em27_metadata.types.SensorMetadata(
                 sensor_id="sid1",
                 serial_number=51,
-                setups=[
-                    src.em27_metadata.types.SetupsListItem(
-                        from_datetime="2020-02-01T01:00:00+0000",# pyright: ignore[reportArgumentType]
-                        to_datetime="2020-02-01T09:59:59+0000",# pyright: ignore[reportArgumentType]
-                        value=src.em27_metadata.types.Setup(
-                            lid="lid1", pds="A", utc_offset=3.7, profile_lid="lid2" # pyright: ignore[reportCallIssue]
-                        ),
+                deployments=[
+                    src.em27_metadata.types.Deployment(
+                        from_datetime="2020-02-01T01:00:00+0000",
+                        to_datetime="2020-02-01T09:59:59+0000",
+                        location_id="lid1",
+                        pressure_data_source="A",
+                        utc_offset=3.7,
+                        atmospheric_profile_location_id="lid2",
                     ),
-                    src.em27_metadata.types.SetupsListItem(
-                        from_datetime="2020-02-01T12:00:00+0000",# pyright: ignore[reportArgumentType]
-                        to_datetime="2020-02-01T21:59:59+0000",# pyright: ignore[reportArgumentType]
-                        value=src.em27_metadata.types.Setup(lid="lid1", pds="B"),# pyright: ignore[reportCallIssue]
+                    src.em27_metadata.types.Deployment(
+                        from_datetime="2020-02-01T12:00:00+0000",
+                        to_datetime="2020-02-01T21:59:59+0000",
+                        location_id="lid1",
+                        pressure_data_source="B",
                     ),
-                    src.em27_metadata.types.SetupsListItem(
-                        from_datetime="2020-02-01T22:00:00+0000",# pyright: ignore[reportArgumentType]
-                        to_datetime="2020-02-03T22:59:59+0000", # pyright: ignore[reportArgumentType]
-                        value=src.em27_metadata.types.Setup(lid="lid1", pds="C"), # pyright: ignore[reportCallIssue]
+                    src.em27_metadata.types.Deployment(
+                        from_datetime="2020-02-01T22:00:00+0000",
+                        to_datetime="2020-02-03T22:59:59+0000",
+                        location_id="lid1",
+                        pressure_data_source="C",
                     ),
-                    src.em27_metadata.types.SetupsListItem(
-                        from_datetime="2020-02-04T00:00:00+0000",# pyright: ignore[reportArgumentType]
-                        to_datetime="2020-02-04T20:59:59+0000",# pyright: ignore[reportArgumentType]
-                        value=src.em27_metadata.types.Setup(lid="lid1", pds="D"),# pyright: ignore[reportCallIssue]
+                    src.em27_metadata.types.Deployment(
+                        from_datetime="2020-02-04T00:00:00+0000",
+                        to_datetime="2020-02-04T20:59:59+0000",
+                        location_id="lid1",
+                        pressure_data_source="D",
                     ),
-                    src.em27_metadata.types.SetupsListItem(
-                        from_datetime="2020-02-04T22:00:00+0000",# pyright: ignore[reportArgumentType]
-                        to_datetime="2020-02-04T23:59:59+0000",# pyright: ignore[reportArgumentType]
-                        value=src.em27_metadata.types.Setup(lid="lid1", pds="E"),# pyright: ignore[reportCallIssue]
+                    src.em27_metadata.types.Deployment(
+                        from_datetime="2020-02-04T22:00:00+0000",
[…]
