+++ src/em27_metadata/interfaces.py
@@ -50,9 +50,9 @@ class EM27MetadataInterface:
 
         # reference existence in sensors.json
         for s1 in sensors.root:
-            for l1 in s1.setups:
-                assert l1.value.location_id in locations.location_ids, (
-                    f"unknown location id {l1.value.location_id}"
+            for l1 in s1.deployments:
+                assert l1.location_id in locations.location_ids, (
+                    f"unknown location id {l1.location_id}"
                 )
 
         # reference existence in campaigns.json
@@ -76,8 +76,8 @@ class EM27MetadataInterface:
         """For a given `sensor_id`, return the list of metadata contexts between
         `from_datetime` and `to_datetime`.
 
-        Each "context" is a time period where the setup is constant. For example,
-        when requesting a full 24 hour day, and the setup changed at noon, the
+        Each "context" is a time period where the deployment is constant. For example,
+        when requesting a full 24 hour day, and the deployment changed at noon, the
         returned list will contain two items: One context until noon, and one
         context after noon.
 
@@ -100,81 +100,86 @@ class EM27MetadataInterface:
         if from_datetime > to_datetime:
             raise ValueError(f"from_datetime ({from_datetime}) > to_datetime ({to_datetime})")
 
-        # find all relevant setups
+        # find all relevant deployments
 
-        setup_spans: list[
+        deployment_spans: list[
             tuple[
-                em27_metadata_types.SetupsListItem,
+                em27_metadata_types.Deployment,
                 datetime.datetime,
                 datetime.datetime,
             ]
         ] = []
-        for setup in sensor.setups:
-            setup_span = (
-                setup,
-                setup.from_datetime_parsed,
-                setup.to_datetime_parsed,
+        for deployment in sensor.deployments:
+            deployment_span = (
+                deployment,
+                deployment.from_datetime_parsed,
+                deployment.to_datetime_parsed,
             )
-            if len(setup_spans) > 0:
-                last_setup, last_from_datetime, last_to_datetime = setup_spans[-1]
+            if len(deployment_spans) > 0:
+                last_deployment, last_from_datetime, last_to_datetime = deployment_spans[-1]
                 if (
-                    (setup_span[1] - last_to_datetime).total_seconds() == 1
-                    and last_setup.value == setup.value
+                    (deployment_span[1] - last_to_datetime).total_seconds() == 1
+                    and last_deployment.model_dump(exclude={"from_datetime", "to_datetime"})
+                    == deployment.model_dump(exclude={"from_datetime", "to_datetime"})
                 ):
-                    setup_spans[-1] = (last_setup, last_from_datetime, setup_span[2])
+                    deployment_spans[-1] = (
+                        last_deployment,
+                        last_from_datetime,
+                        deployment_span[2],
+                    )
                     continue
-            setup_spans.append(setup_span)
+            deployment_spans.append(deployment_span)
 
-        relevant_setups = [
+        relevant_deployments = [
             (
-                setup,
-                setup_from_datetime.astimezone(datetime.timezone.utc),
-                setup_to_datetime.astimezone(datetime.timezone.utc),
+                deployment,
+                deployment_from_datetime.astimezone(datetime.timezone.utc),
+                deployment_to_datetime.astimezone(datetime.timezone.utc),
             )
-            for setup, setup_from_datetime, setup_to_datetime in setup_spans
+            for deployment, deployment_from_datetime, deployment_to_datetime in deployment_spans
             if tum_esm_utils.timing.datetime_span_intersection(
                 (from_datetime, to_datetime),
-                (setup_from_datetime, setup_to_datetime),
+                (deployment_from_datetime, deployment_to_datetime),
             )
             is not None
         ]
 
-        for s1, s2 in zip(relevant_setups[:-1], relevant_setups[1:]):
[…]
