### 2026-09-12 9b6b0c4
#155 Integrate metadata library into this repository and also migrate to TOML (14)
Add test whether full backwards compatibility of the old metadata format and old usage of the metadata library is intact.
