### 2026-09-13 2bb9cc4
#155 Integrate metadata library into this repository and also migrate to TOML (16)
Add CLI command to fetch remote metadata and store in locally in the new TOML format
