### 2026-09-13 4a52cfe
#155 Integrate metadata library into this repository and also migrate to TOML (17)
Add package distribution environment
