### 2026-09-13 de8b37f
#155 Integrate metadata library into this repository and also migrate to TOML (18)
Add more rigorous tests for full backwards compatibility of the library
