### 2026-09-13 f5061f4
#155 Integrate metadata library into this repository and also migrate to TOML (19)
Add a new script to publish the built product to the PyPI registry
