### 2026-09-13 e318286
#155 Integrate metadata library into this repository and also migrate to TOML (20)
Update a few small things
