### 2026-09-13 f62abce
#155 Integrate metadata library into this repository and also migrate to TOML (21)
Add GitHub action to this repository
