+++ .github/actions/validate-metadata/action.yml
@@ -1,0 +1,33 @@
+name: "Validate EM27/SUN metadata"
+author: "Moritz Oliveira Makowski <moritz.makowski@tum.de>"
+description: "Validate an EM27/SUN metadata-storage repository using either the TOML or legacy JSON schema"
+
+inputs:
+  metadata-directory:
+    description: "Directory containing em27_metadata.toml or the legacy data/*.json files"
+    required: false
+    default: "."
+
+runs:
+  using: "composite"
+  steps:
+    - name: Set up Python 3.14
+      uses: actions/setup-python@v5
+      with:
+        python-version: "3.14"
+
+    - name: Install validation dependencies
+      shell: bash
+      run: >-
+        python -m pip install
+        "pydantic>=2.10.2"
+        "tomli>=2.4.0"
+        "tomli-w>=1.2.0"
+        "tum-esm-utils>=2.11.3"
+
+    - name: Validate metadata
+      shell: bash
+      env:
+        EM27_METADATA_DIRECTORY: ${{ inputs.metadata-directory }}
+        EM27_METADATA_SOURCE_DIR: ${{ github.action_path }}/../../../src
+      run: python "${{ github.action_path }}/validate.py"
