### 2026-09-13 c078872
Refine docs (20)
Update docs sidebar

Improve name of new metadata migration CLI command

Point README and index page to quick start guide

Move getting started section further up in the README
