+++ README.md
@@ -19,8 +19,6 @@ This codebase provides an automated data pipeline for [Proffast 1 and 2.X](https
 
 This pipeline does not intend to replace the Proffast Pylot by any means. It uses most of the Pylot's functionality and provides an automation layer on top of that. You can use Proffast directly for full flexibility, the Pylot for some flexibility, or this pipeline, which is more opinionated than the Pylot but highly automated. Furthermore, this pipeline will continue to include support for the different retrieval algorithms.
 
-We decided to include a copy of the Python codebase of the Proffast Pylot inside this repository so we can have less complexity due to Git Submodules or on-demand downloads. Since the Proffast Pylot only supports Proffast 2.X, we have added our own wrapper for Proffast 1 in this pipeline.
-
 ## What does this pipeline do?
 
 The pipeline consists of three building blocks that are required to retrieve EM27 data:
@@ -42,7 +40,11 @@ The data flow from input to merged outputs:
   className="my-8 max-w-lg mx-auto w-full p-2 md:p-4 bg-white rounded"
 />
 
-## Features
+## Getting started
+
+Please visit the [Quick Start Guide](https://tum-esm.github.io/em27-retrieval-pipeline/guides/quick-start) to install the pipeline on your system and run the example. You can look into the API Reference pages for an exact schema definition and explanation of every parameter. You can also use this documentation's search function (top right) to find what you are looking for.
+
+## All features of this pipeline
 
 **Easy configuration of using a validated `config.json` (and metadata files):** By "validated", we mean that before the processing starts, the config files content will be parsed and validated against a JSON schema. This way, you can be sure that the pipeline will not fail due to a misconfiguration, and you will immediately get precise error messages.
 
@@ -60,22 +62,10 @@ The data flow from input to merged outputs:
 
 **Documentation and complete API reference:** hosted at [em27-retrieval-pipeline.netlify.app](https://tum-esm.github.io/em27-retrieval-pipeline/)
 
-## Getting Started
-
-To fully understand the pipeline, you should read the following sections of the documentation:
-
-1. The [configuration section](https://tum-esm.github.io/em27-retrieval-pipeline/guides/configuration) explains how the pipeline uses a `config.json` file to read the parameters of your environment.
-2. The [directories section](https://tum-esm.github.io/em27-retrieval-pipeline/guides/directories) explains how the input directories should be structured and how the pipeline structures its outputs.
-3. The [metadata section](https://tum-esm.github.io/em27-retrieval-pipeline/guides/metadata) explains what the metadata the pipeline requires is structured and how to connect it to the pipeline.
-4. The [installation section](https://tum-esm.github.io/em27-retrieval-pipeline/guides/installation) explains how to install the pipeline.
-5. The [usage section](https://tum-esm.github.io/em27-retrieval-pipeline/guides/usage) explains how to run the pipeline.
-
-The `example` directory of this repository contains a fully working example of the different steps of the pipeline.
-
-You can look into the API Reference pages for an exact schema definition and explanation of every parameter. You can also use this documentation's search function (top right) to find what you are looking for.
-
 ## Contributing
 
 We are happy to include any feature that is useful for the COCCON community. We are also open to expand the pipeline to other retrieval algorithms or spectrometers.
 
[…]
