### 2026-09-13 5215c52
Refine docs (21)
Clean up and refine the config and geoms metadata schemas and docs
