+++ src/types/config.py
@@ -50,22 +50,6 @@ class MetadataConfig(pydantic.BaseModel):
 
 
 class DataSubConfigs:
-    class AtmosphericProfiles(pydantic.BaseModel):
-        """Where to find the atmospheric profile files."""
-
-        path: tum_esm_utils.validators.StrictDirectoryPath = pydantic.Field(
-            ...,
-            description="Directory path to atmospheric profile files.",
-        )
-
-        # before the validation, if the path is relative, make it absolute based on the cwd
-        @pydantic.model_validator(mode="before")
-        def _make_path_absolute(cls, values: Any) -> Any:
-            assert isinstance(values["path"], str)
-            if not os.path.isabs(values["path"]):
-                values["path"] = os.path.abspath(values["path"])
-            return values
-
     class GroundPressure(pydantic.BaseModel):
         """Format of the ground pressure files. We support any text file that stores one data point per row and separates the columns with a comma, space, or tab, i.e. CSV, TSV, or space-separated files. Using the `file_regex` field, you specify which files to consider for a given sensor id and date.
 
@@ -80,7 +64,7 @@ class DataSubConfigs:
         # where to find the files
         path: tum_esm_utils.validators.StrictDirectoryPath = pydantic.Field(
             ...,
-            description="Directory path to ground pressure files.",
+            description="Directory path to ground pressure files. You should use absolute paths, but if you need relative paths, then this is relative to the caller of the CLI or the pipeline's entrypoint.",
         )
 
         # how to find and parse the files
@@ -208,7 +192,7 @@ class DataSubConfigs:
 
         path: tum_esm_utils.validators.StrictDirectoryPath = pydantic.Field(
             ...,
-            description="Directory path to atmospheric profile files.",
+            description="Directory path to atmospheric profile files. You should use absolute paths, but if you need relative paths, then this is relative to the caller of the CLI or the pipeline's entrypoint.",
         )
         ifg_file_regex: str = pydantic.Field(
             ...,
@@ -228,12 +212,28 @@ class DataSubConfigs:
                 values["path"] = os.path.abspath(values["path"])
             return values
 
+    class AtmosphericProfiles(pydantic.BaseModel):
+        """Where to find the atmospheric profile files."""
+
+        path: tum_esm_utils.validators.StrictDirectoryPath = pydantic.Field(
+            ...,
+            description="Directory path to atmospheric profile files. You should use absolute paths, but if you need relative paths, then this is relative to the caller of the CLI or the pipeline's entrypoint.",
+        )
+
+        # before the validation, if the path is relative, make it absolute based on the cwd
+        @pydantic.model_validator(mode="before")
+        def _make_path_absolute(cls, values: Any) -> Any:
+            assert isinstance(values["path"], str)
+            if not os.path.isabs(values["path"]):
+                values["path"] = os.path.abspath(values["path"])
+            return values
+
     class Results(pydantic.BaseModel):
         """Where to find the results."""
 
         path: tum_esm_utils.validators.StrictDirectoryPath = pydantic.Field(
             ...,
-            description="Directory path to atmospheric profile files.",
+            description="Directory path to atmospheric profile files. You should use absolute paths, but if you need relative paths, then this is relative to the caller of the CLI or the pipeline's entrypoint.",
         )
 
         # before the validation, if the path is relative, make it absolute based on the cwd
@@ -269,13 +269,15 @@ class GGGProfilesDownloaderSubConfigs:
             description="Email address to use to log in to the ccycle ftp server.",
         )
         max_parallel_requests: int = pydantic.Field(
-            ...,
+            default=25,
             ge=1,
             le=200,
[…]
