+++ README.md
@@ -42,11 +42,11 @@ The data flow from input to merged outputs:
 
 ## Getting started
 
-Please visit the [Quick Start Guide](https://tum-esm.github.io/em27-retrieval-pipeline/guides/quick-start) to install the pipeline on your system and run the example. You can look into the API Reference pages for an exact schema definition and explanation of every parameter. You can also use this documentation's search function (top right) to find what you are looking for.
+Please visit the [Quick Start Guide](https://tum-esm.github.io/em27-retrieval-pipeline/guides/quick-start) to install the pipeline on your system and run the example. You can look into the API Reference pages for an exact schema definition and explanation of every parameter. You can also use the documentation's search function (top navbar) to find what you are looking for.
 
 ## All features of this pipeline
 
-**Easy configuration of using a validated `config.json` (and metadata files):** By "validated", we mean that before the processing starts, the config files content will be parsed and validated against a JSON schema. This way, you can be sure that the pipeline will not fail due to a misconfiguration, and you will immediately get precise error messages.
+**Easy configuration of using a validated `config.toml` (and metadata files):** By "validated", we mean that before the processing starts, the config files content will be parsed and validated against a TOML schema. This way, you can be sure that the pipeline will not fail due to a misconfiguration, and you will immediately get precise error messages.
 
 **Opinionated management of station metadata:** We manage our EM27 metadata using JSON files instead of database tables, which has several benefits mentioned in the [metadata repository](https://github.com/tum-esm/em27-metadata)
 
