### 2026-09-13 ddc5c67
Refine docs (22)
Add reference page for the EM27 metadata
