### 2026-09-13 6dfc9d4
Refine docs (23)
Refine CLI reference page: group commands into sections and add dividers
