### 2026-09-13 7440b83
Migrate all configs used in example to TOML
