+++ example/config/config.toml
@@ -1,0 +1,77 @@
+version = "1.11"
+
+# METADATA
+
+[metadata]
+source = "local"
+
+# DATA
+
+[data.atmospheric_profiles]
+path = "data/inputs/atmospheric-profiles"
+
+[data.ground_pressure]
+path = "data/inputs/ground-pressure"
+file_regex = "^ground-pressure-$(SENSOR_ID)-$(YYYY)-$(MM)-$(DD).csv$"
+separator = ","
+date_column = "utc-date"
+date_column_format = "%Y-%m-%d"
+time_column = "utc-time"
+time_column_format = "%H:%M:%S"
+pressure_column = "pressure"
+pressure_column_format = "hPa"
+
+[data.interferograms]
+path = "data/inputs/interferograms"
+ifg_file_regex = "^$(SENSOR_ID)$(DATE).*\\.\\d+$"
+
+[data.results]
+path = "data/outputs/individual"
+
+# RETRIEVAL
+
+[retrieval.general]
+max_process_count = 8
+queue_verbosity = "compact"
+
+[retrieval.jobs.0]
+retrieval_algorithm = "proffast-2.4"
+atmospheric_profile_model = "GGG2020"
+sensor_ids = [ "so", "mc"]
+from_date = "2017-01-01"
+to_date = "2022-12-31"
+store_binary_spectra = false
+dc_min_threshold = 0.05
+dc_var_threshold = 0.1
+use_local_pressure_in_pcxs = false
+use_ifg_corruption_filter = true
+
+# BUNDLE EXPORTS
+
+[[bundle_exports]]
+dst_dir = "data/outputs/bundles"
+output_formats = [ "csv", "parquet" ]
+from_datetime = "2017-01-01T00:00:00Z"
+to_datetime = "2022-12-31T23:59:59Z"
+retrieval_algorithms = ["proffast-2.4" ]
+atmospheric_profile_models = [ "GGG2020" ]
+sensor_ids = [ "so", "mc" ]
+parse_dc_timeseries = true
+parse_retrieval_diagnostics = true
+
+# GEOMS EXPORTS
+
+[[geoms_exports]]
+sensor_ids = [ "so", "mc" ]
+retrieval_algorithms = [ "proffast-2.4" ]
+atmospheric_profile_models = [ "GGG2020" ]
+from_datetime = "2017-01-01T00:00:00Z"
+to_datetime = "2022-12-31T23:59:59Z"
+parse_dc_timeseries = true
+dc_min_xco2 = 0.05
+dc_min_xch4 = 0.05
+dc_min_xh2o = 0.05
+dc_min_xco = 0.05
+max_sza = 80.0
+min_xair = 0.98
+max_xair = 1.02
