+++ src/em27_metadata/loader.py
@@ -170,9 +170,7 @@ def load_from_local_files(
         if os.path.isfile(toml_path):
             with open(toml_path, "rb") as file:
                 toml_data = tomli.load(file)
-            em27_metadata_object = em27_metadata_types.EM27MetadataObject.model_validate(
-                toml_data
-            )
+            em27_metadata_object = em27_metadata_types.EM27MetadataObject.model_validate(toml_data)
             return em27_metadata_interfaces.EM27MetadataInterface(
                 locations=em27_metadata_object.locations,
                 sensors=em27_metadata_object.sensors,
