+++ src/types/basic_types.py
@@ -3,9 +3,7 @@ from typing import Literal
 
 
 DATE_STRING_PATTERN = r"^\d{4}-\d{2}-\d{2}$"
-UTC_DATETIME_STRING_PATTERN = (
-    r"^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:Z|\+0000)$"
-)
+UTC_DATETIME_STRING_PATTERN = r"^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:Z|\+0000)$"
 
 
 def parse_date_string(value: str) -> datetime.date:
@@ -14,6 +12,7 @@ def parse_date_string(value: str) -> datetime.date:
 
 def parse_datetime_string(value: str) -> datetime.datetime:
     return datetime.datetime.strptime(value, "%Y-%m-%dT%H:%M:%S%z")
+
 
 RetrievalAlgorithm = Literal[
     "proffast-1.0",
