### 2026-09-13 8aa9de8
Fix ruff linting issues
