+++ src/types/old_schemas.py
@@ -1,12 +1,14 @@
 from __future__ import annotations
-from typing import Any, Literal, Optional
-import os
+
 import datetime
+import os
+from typing import Any, Literal, Optional
+
 import dotenv
 import pydantic
 import tum_esm_utils
-from .basic_types import AtmosphericProfileModel, RetrievalAlgorithm
 
+from .basic_types import AtmosphericProfileModel, RetrievalAlgorithm
 
 # CONFIG
 
