+++ src/utils/functions.py
@@ -1,6 +1,7 @@
 import datetime
 import os
 from typing import Any, Optional
+
 import polars as pl
 import tomli
 import tum_esm_utils
