+++ src/bundle_exporter/main.py
@@ -9,7 +9,7 @@ import tqdm
 import tum_esm_utils
 
 sys.path.append(tum_esm_utils.files.rel_to_abs_path("../.."))
-from src import types, utils, em27_metadata
+from src import em27_metadata, types, utils
 
 from .load_results import load_results_directory
 
