### 2026-09-13 bafa011
Exclude some artifacts when building the em27_metadata library
