### 2026-09-13 47b496e
Expand version number tests and debug docs sync test
