+++ tests/repository/test_docs_sync.py
@@ -1,4 +1,5 @@
 import sys
+
 import pytest
 import tum_esm_utils
 
@@ -8,7 +9,7 @@ _PROJECT_DIR = tum_esm_utils.files.rel_to_abs_path("../..")
 def _get_checksum() -> str:
     # credits to https://stackoverflow.com/a/545413/8255842
     return tum_esm_utils.shell.run_shell_command(
-        "find docs/pages docs/components -type f -print0 | sort -z | xargs -0 sha1sum | sha1sum",
+        "find docs/src -type f -print0 | sort -z | xargs -0 sha1sum | sha1sum",
         working_directory=_PROJECT_DIR,
     )
 
