+++ packaging/em27-metadata/pyproject.toml
@@ -11,12 +11,12 @@ dependencies = [
     "tomli>=2.4.0",
     "tum-esm-utils>=2.11.3",
 ]
-requires-python = ">=3.10,<4.0"
+requires-python = ">=3.11,<4.0"
 readme = "README.md"
-license = { text = "GPL-3.0" }
+license = { text = "GPL-3.0-or-later" }
 classifiers = [
+    "License :: OSI Approved :: GNU General Public License (GPL)",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
