### 2026-09-13 93de13a
Rename CLI command group "profiles" to "ggg-profiles-downloader"
Calling "cli.py profiles" continues to work to ensure backwards compatibility
