+++ README.md
@@ -58,7 +58,7 @@ Please visit the [Quick Start Guide](https://tum-esm.github.io/em27-retrieval-pi
 
 **Comprehensive logs and output data management:** It will store failed and succeeded containers. The output is the same as with the Pylot but also contains all config files the pipeline used to run this container and logs generated by the container.
 
-**Bundling of retrieval results:** The raw retrieval outputs will be distributed over hundreds or thousands of folders and files. The bundling script of this pipeline will merge these outputs into one file per station, retrieval algorithm, and atmospheric profile. This way, you can easily access the data for further processing. ([Read about it here](/guides/data-directories#bundles)).
+**Bundling of retrieval results:** The raw retrieval outputs will be distributed over hundreds or thousands of folders and files. The bundling script of this pipeline will merge these outputs into one file per station, retrieval algorithm, and atmospheric profile. This way, you can easily access the data for further processing. Read about it ([here](https://tum-esm.github.io/em27-retrieval-pipeline/guides/data-directories#bundles)).
 
 **Documentation and complete API reference:** hosted at [tum-esm.github.io/em27-retrieval-pipeline](https://tum-esm.github.io/em27-retrieval-pipeline/)
 
