### 2026-09-13 e5cd44d
Clean up code on how the metadata is loaded
