+++ tests/integration/test_em27_metadata_connection.py
@@ -7,12 +7,9 @@ import src
 @pytest.mark.metadata
 def test_metadata_connection() -> None:
     config = src.types.Config.load(ignore_path_existence=True)
-    # TODO: refactor metadata loading
-    if config.metadata.source == "local":
-        src.em27_metadata.load_from_local_files(config_directory=src.types.Config.get_config_dir())
-    else:
-        assert config.metadata.github_repository is not None, "This should not happen"
-        src.em27_metadata.load_from_github(
-            github_repository=config.metadata.github_repository,
-            access_token=config.metadata.github_access_token,
-        )
+    src.em27_metadata.load_from_config(
+        source=config.metadata.source,
+        github_repository=config.metadata.github_repository,
+        github_access_token=config.metadata.github_access_token,
+        config_directory=src.types.Config.get_config_dir(),
+    )
