+++ tests/repository/test_local_metadata.py
@@ -38,26 +38,57 @@ def store_local_metadata_during_test() -> Generator[None, None, None]:
 @pytest.mark.order(3)
 @pytest.mark.quick
 def test_local_metadata_loader(store_local_metadata_during_test: None) -> None:
-    assert src.utils.metadata.load_local_em27_metadata_interface(config_dir=CONFIG_DIR) is None
+    with pytest.raises(FileNotFoundError):
+        src.em27_metadata.load_from_config(
+            source="local",
+            github_repository=None,
+            github_access_token=None,
+            config_directory=CONFIG_DIR,
+        )
 
     with open(LOCATIONS_PATH, "w") as f:
         f.write("[]")
     with pytest.raises(FileNotFoundError):
-        src.utils.metadata.load_local_em27_metadata_interface(config_dir=CONFIG_DIR)
+        src.em27_metadata.load_from_config(
+            source="local",
+            github_repository=None,
+            github_access_token=None,
+            config_directory=CONFIG_DIR,
+        )
 
     with open(SENSORS_PATH, "w") as f:
         f.write("[]")
-    src.utils.metadata.load_local_em27_metadata_interface(config_dir=CONFIG_DIR)
+    src.em27_metadata.load_from_config(
+        source="local",
+        github_repository=None,
+        github_access_token=None,
+        config_directory=CONFIG_DIR,
+    )
 
     with open(CAMPAIGNS_PATH, "w") as f:
         f.write("[]")
-    src.utils.metadata.load_local_em27_metadata_interface(config_dir=CONFIG_DIR)
+    src.em27_metadata.load_from_config(
+        source="local",
+        github_repository=None,
+        github_access_token=None,
+        config_directory=CONFIG_DIR,
+    )
 
     with open(EVENTS_PATH, "w") as f:
         f.write("[]")
-    src.utils.metadata.load_local_em27_metadata_interface(config_dir=CONFIG_DIR)
+    src.em27_metadata.load_from_config(
+        source="local",
+        github_repository=None,
+        github_access_token=None,
+        config_directory=CONFIG_DIR,
+    )
 
-    interface = src.utils.metadata.load_local_em27_metadata_interface(config_dir=CONFIG_DIR)
+    interface = src.em27_metadata.load_from_config(
+        source="local",
+        github_repository=None,
+        github_access_token=None,
+        config_directory=CONFIG_DIR,
+    )
     assert interface is not None
     assert len(interface.locations.root) == 0
     assert len(interface.sensors.root) == 0
