+++ src/utils/__init__.py
@@ -1,5 +1,4 @@
 from . import functions as functions
-from . import metadata as metadata
 from . import report as report
 from . import semaphores as semaphores
 from . import text as text
