+++ src/em27_metadata/__init__.py
@@ -5,6 +5,7 @@ Retrieval Pipeline repository since the pipeline is the main consumer of it."""
 
 from . import types as types
 from .interfaces import EM27MetadataInterface as EM27MetadataInterface
+from .loader import load_from_config as load_from_config
 from .loader import load_from_example_data as load_from_example_data
 from .loader import load_from_github as load_from_github
 from .loader import load_from_local_files as load_from_local_files
