### 2026-09-13 1194412
Make output of automatic json-to-toml migration prettier
Now the keys are ordered, empty objects are omitted, and the generated TOML files contain the dividing comments
