+++ src/utils/__init__.py
@@ -2,3 +2,4 @@ from . import functions as functions
 from . import report as report
 from . import semaphores as semaphores
 from . import text as text
+from . import toml as toml
