+++ tests/repository/test_config_templates.py
@@ -5,8 +5,8 @@ import tempfile
 import pytest
 import tomli
 import tum_esm_utils
-import src
 
+import src
 
 _OLD_TEMPLATES_DIR = tum_esm_utils.files.rel_to_abs_path("../../config/old_templates")
 
@@ -19,6 +19,16 @@ def _copy_old_templates(tmp_dir: str, filenames: list[str]) -> None:
 def _load_toml_file(filepath: str) -> dict[str, object]:
     with open(filepath, "rb") as file:
         return tomli.load(file)
+
+
+def _load_text_file(filepath: str) -> str:
+    with open(filepath, "r", encoding="utf-8") as file:
+        return file.read()
+
+
+def _assert_in_order(text: str, values: list[str]) -> None:
+    positions = [text.index(value) for value in values]
+    assert positions == sorted(positions)
 
 
 @pytest.mark.quick
@@ -67,6 +77,22 @@ def test_automatic_config_conversion() -> None:
             context={"ignore-path-existence": True},
         )
         assert converted_config == expected_config
+        converted_text = _load_text_file(os.path.join(tmp_dir, "config.toml"))
+        _assert_in_order(
+            converted_text,
+            [
+                "# METADATA",
+                "# DATA",
+                "# GGG PROFILES DOWNLOADER",
+                "# RETRIEVAL",
+                "# BUNDLE GENERATOR",
+                "# GEOMS GENERATOR",
+            ],
+        )
+        assert " = [\n" not in converted_text
+        assert "[retrieval.jobs.1.custom_ils]" not in converted_text
+        assert "[retrieval.jobs.1.pressure_calibration_factors]" not in converted_text
+        assert "[retrieval.jobs.1.pressure_calibration_offsets]" not in converted_text
 
 
 @pytest.mark.order(2)
@@ -93,6 +119,15 @@ def test_automatic_em27_metadata_conversion() -> None:
             os.path.join(_OLD_TEMPLATES_DIR, "em27_metadata.automatically_converted.toml")
         )
         assert converted_metadata == expected_metadata
+        converted_text = _load_text_file(os.path.join(tmp_dir, "em27_metadata.toml"))
+        _assert_in_order(
+            converted_text,
+            ["# LOCATIONS", "# SENSORS", "# CAMPAIGNS", "# EVENTS"],
+        )
+        assert "[[sensors.deployments]]" not in converted_text
+        assert 'deployments = [\n    { from_datetime = "' in converted_text
+        assert 'sensor_ids = [ "so", "mc" ]' in converted_text
+        assert 'location_ids = [ "SOD", "ZEN" ]' in converted_text
 
 
 @pytest.mark.order(2)
@@ -115,3 +150,7 @@ def test_automatic_geoms_metadata_conversion(monkeypatch: pytest.MonkeyPatch) ->
             os.path.join(_OLD_TEMPLATES_DIR, "geoms_metadata.automatically_converted.toml")
         )
         assert converted_metadata == expected_metadata
+        _assert_in_order(
+            _load_text_file(os.path.join(tmp_dir, "geoms_metadata.toml")),
+            ["# METADATA", "# CALIBRATION FACTORS"],
+        )
