### 2026-09-13 9d761bb
Fix warnings and formatting issues
