+++ packaging/em27-metadata/pyproject.toml
@@ -15,7 +15,6 @@ requires-python = ">=3.11,<4.0"
 readme = "README.md"
 license = { text = "GPL-3.0-or-later" }
 classifiers = [
-    "License :: OSI Approved :: GNU General Public License (GPL)",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
