### 2026-09-13 4b79954
Possibly fix the permission bug when using the metadata validation action
