+++ src/em27_metadata/__init__.py
@@ -3,6 +3,7 @@ working with EM27 metadata. It used to be hosted in its own repository at
 https://github.com/tum-esm/em27-metadata but is now included in this EM27
 Retrieval Pipeline repository since the pipeline is the main consumer of it."""
 
+from . import toml as toml
 from . import types as types
 from .interfaces import EM27MetadataInterface as EM27MetadataInterface
 from .loader import load_from_config as load_from_config
