+++ src/em27_metadata/loader.py
@@ -4,8 +4,8 @@ from typing import Callable, Literal, Optional
 import tomli
 import tum_esm_utils
 
-from ..utils import toml as toml_utils
 from . import interfaces as em27_metadata_interfaces
+from . import toml as toml_utils
 from . import types as em27_metadata_types
 
 
@@ -287,8 +287,8 @@ def load_from_local_files(
         toml_utils.dump_pretty_toml_file(
             filepath=tmp_toml_path,
             data=em27_metadata_object.model_dump(mode="json", exclude_none=True),
-            template_filepath=tum_esm_utils.files.rel_to_abs_path(
-                "../../config/em27_metadata.template.toml"
+            template_filepath=os.path.join(
+                os.path.dirname(__file__), "em27_metadata.template.toml"
             ),
             inline_sensor_deployments=True,
             inline_metadata_id_lists=True,
