+++ packaging/em27-metadata/pyproject.toml
@@ -9,6 +9,7 @@ authors = [
 dependencies = [
     "pydantic>=2.12.3",
     "tomli>=2.4.0",
+    "tomli-w>=1.2.0",
     "tum-esm-utils>=2.11.3",
 ]
 requires-python = ">=3.11,<4.0"
