+++ src/em27_metadata/em27_metadata.template.toml
@@ -1,0 +1,15 @@
+# LOCATIONS
+
+locations = []
+
+# SENSORS
+
+sensors = []
+
+# CAMPAIGNS
+
+campaigns = []
+
+# EVENTS
+
+events = []
