### 2026-09-13 e9ab264
Make the README of the em27-metadata library pretty again
