### 2026-09-13 0789575
Debug tests of the new em27 metadata library bundling
