+++ CITATION.cff
@@ -39,5 +39,5 @@ doi: 10.5281/zenodo.14284968
 license:
   - gpl-3.0-or-later
 type: software
-version: 1.11.0
+version: 1.11.1
 date-released: "2026-09-13"
