+++ packaging/em27-metadata/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "em27_metadata"
-version = "1.11.0"
+version = "1.11.1"
 description = "Single source of truth for ESM's EM27 measurement logistics"
 authors = [
     { name = "Moritz Oliveira Makowski", email = "moritz.makowski@tum.de" },
